parser/tests/tree_construction/passing.dat

#data
<!DOCTYPE html><html><head></head><body><b>JAKE</b></body></html>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       "JAKE"

#data
Test
#errors
(1,0): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "Test"

#data
<p>One<p>Two
#errors
(1,3): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <p>
|       "One"
|     <p>
|       "Two"

#data
Line1<br>Line2<br>Line3<br>Line4
#errors
(1,0): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "Line1"
|     <br>
|     "Line2"
|     <br>
|     "Line3"
|     <br>
|     "Line4"

#data
<html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<head>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<body>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head></head>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head></head><body>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head></head><body></body>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head><body></body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head></body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><head><body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<html><body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
<head></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>

#data
</head>
#errors
(1,7): expected-doctype-but-got-end-tag
#document
| <html>
|   <head>
|   <body>

#data
</body>
#errors
(1,7): expected-doctype-but-got-end-tag element.
#document
| <html>
|   <head>
|   <body>

#data
</html>
#errors
(1,7): expected-doctype-but-got-end-tag element.
#document
| <html>
|   <head>
|   <body>

#data
<b><table><td><i></table>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,25): unexpected-cell-end-tag
(1,25): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <i>

#data
<b><table><td></b><i></table>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,18): unexpected-end-tag
(1,29): unexpected-cell-end-tag
(1,30): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <i>
|       "X"

#data
<h1>Hello<h2>World
#errors
(1,4): expected-doctype-but-got-start-tag
(1,13): unexpected-start-tag
(1,18): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <h1>
|       "Hello"
|     <h2>
|       "World"

#data
<a><p>X<a>Y</a>Z</p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,10): unexpected-start-tag-implies-end-tag
(1,10): adoption-agency-1.3
(1,24): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <p>
|       <a>
|         "X"
|       <a>
|         "Y"
|       "Z"

#data
<b><button>foo</b>bar
#errors
(1,3): expected-doctype-but-got-start-tag
(1,18): adoption-agency-1.3
(1,21): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|     <button>
|       <b>
|         "foo"
|       "bar"

#data
<!DOCTYPE html><span><button>foo</span>bar
#errors
(1,39): unexpected-end-tag
(1,42): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <span>
|       <button>
|         "foobar"

#data
<p><b><div><marquee></p></b></div>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag
(1,24): unexpected-end-tag
(1,28): unexpected-end-tag
(1,34): end-tag-too-early
(1,35): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <div>
|       <b>
|         <marquee>
|           <p>
|           "X"

#data
<script><div></script></div><title><p></title><p><p>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,28): unexpected-end-tag
#document
| <html>
|   <head>
|     <script>
|       "<div>"
|     <title>
|       "<p>"
|   <body>
|     <p>
|     <p>

#data
<!--><div>--<!-->
#errors
(1,5): incorrect-comment
(1,10): expected-doctype-but-got-start-tag
(1,17): incorrect-comment
(1,17): expected-closing-tag-but-got-eof
#new-errors
(1:5) abrupt-closing-of-empty-comment
(1:17) abrupt-closing-of-empty-comment
#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     <div>
|       "--"
|       <!--  -->

#data
<p><hr></p>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <hr>
|     <p>

#data
<select><b><option><select><option></b></select>X
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): unexpected-start-tag-in-select
(1,27): unexpected-select-in-select
(1,39): unexpected-end-tag
(1,48): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|     <option>
|       "X"

#data
<a><table><td><a><table></table><a></tr><a></table><b>X</b>C<a>Y
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,35): unexpected-start-tag-implies-end-tag
(1,40): unexpected-cell-end-tag
(1,43): unexpected-start-tag-implies-table-voodoo
(1,43): unexpected-start-tag-implies-end-tag
(1,43): unexpected-end-tag
(1,63): unexpected-start-tag-implies-end-tag
(1,64): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|     <a>
|       <b>
|         "X"
|       "C"
|     <a>
|       "Y"

#data
<a X>0<b>1<a Y>2
#errors
(1,5): expected-doctype-but-got-start-tag
(1,15): unexpected-start-tag-implies-end-tag
(1,15): adoption-agency-1.3
(1,16): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       x=""
|       "0"
|       <b>
|         "1"
|     <b>
|       <a>
|         y=""
|         "2"

#data
<!-----><font><div>hello<table>excite!<b>me!<th><i>please!</tr><!--X-->
#errors
(1,7): unexpected-dash-after-double-dash-in-comment
(1,14): expected-doctype-but-got-start-tag
(1,41): unexpected-start-tag-implies-table-voodoo
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): foster-parenting-character-in-table
(1,48): unexpected-cell-in-table-body
(1,63): unexpected-cell-end-tag
(1,71): eof-in-table
#document
| <!-- - -->
| <html>
|   <head>
|   <body>
|     <font>
|       <div>
|         "helloexcite!"
|         <b>
|           "me!"
|         <table>
|           <tbody>
|             <tr>
|               <th>
|                 <i>
|                   "please!"
|             <!-- X -->

#data
<!DOCTYPE html><li>hello<li>world<ul>how<li>do</ul>you</body><!--do-->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <li>
|       "hello"
|     <li>
|       "world"
|       <ul>
|         "how"
|         <li>
|           "do"
|       "you"
|   <!-- do -->

#data
<!DOCTYPE html>A<option>B<optgroup>C<select>D</option>E
#errors
(1,54): unexpected-end-tag-in-select
(1,55): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "A"
|     <option>
|       "B"
|     <optgroup>
|       "C"
|       <select>
|         "DE"

#data
<
#errors
(1,1): expected-tag-name
(1,1): expected-doctype-but-got-chars
#new-errors
(1:2) eof-before-tag-name
#document
| <html>
|   <head>
|   <body>
|     "<"

#data
<#
#errors
(1,1): expected-tag-name
(1,1): expected-doctype-but-got-chars
#new-errors
(1:2) invalid-first-character-of-tag-name
#document
| <html>
|   <head>
|   <body>
|     "<#"

#data
</
#errors
(1,2): expected-closing-tag-but-got-eof
(1,2): expected-doctype-but-got-chars
#new-errors
(1:3) eof-before-tag-name
#document
| <html>
|   <head>
|   <body>
|     "</"

#data
</#
#errors
(1,2): expected-closing-tag-but-got-char
(1,3): expected-doctype-but-got-eof
#new-errors
(1:3) invalid-first-character-of-tag-name
#document
| <!-- # -->
| <html>
|   <head>
|   <body>

#data
<?
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,2): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ? -->
| <html>
|   <head>
|   <body>

#data
<?#
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,3): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ?# -->
| <html>
|   <head>
|   <body>

#data
<!
#errors
(1,2): expected-dashes-or-doctype
(1,2): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!--  -->
| <html>
|   <head>
|   <body>

#data
<!#
#errors
(1,2): expected-dashes-or-doctype
(1,3): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!-- # -->
| <html>
|   <head>
|   <body>

#data
<?COMMENT?>
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,11): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ?COMMENT? -->
| <html>
|   <head>
|   <body>

#data
<!COMMENT>
#errors
(1,2): expected-dashes-or-doctype
(1,10): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!-- COMMENT -->
| <html>
|   <head>
|   <body>

#data
</ COMMENT >
#errors
(1,2): expected-closing-tag-but-got-char
(1,12): expected-doctype-but-got-eof
#new-errors
(1:3) invalid-first-character-of-tag-name
#document
| <!--  COMMENT  -->
| <html>
|   <head>
|   <body>

#data
<?COM--MENT?>
#errors
(1,1): expected-tag-name-but-got-question-mark
(1,13): expected-doctype-but-got-eof
#new-errors
(1:2) unexpected-question-mark-instead-of-tag-name
#document
| <!-- ?COM--MENT? -->
| <html>
|   <head>
|   <body>

#data
<!COM--MENT>
#errors
(1,2): expected-dashes-or-doctype
(1,12): expected-doctype-but-got-eof
#new-errors
(1:3) incorrectly-opened-comment
#document
| <!-- COM--MENT -->
| <html>
|   <head>
|   <body>

#data
</ COM--MENT >
#errors
(1,2): expected-closing-tag-but-got-char
(1,14): expected-doctype-but-got-eof
#new-errors
(1:3) invalid-first-character-of-tag-name
#document
| <!--  COM--MENT  -->
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><style> EOF
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       " EOF"
|   <body>

#data
<!DOCTYPE html><script> <!-- </script> --> </script> EOF
#errors
(1,52): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       " <!-- "
|     " "
|   <body>
|     "-->  EOF"

#data
<b><p></b>TEST
#errors
(1,3): expected-doctype-but-got-start-tag
(1,10): adoption-agency-1.3
#document
| <html>
|   <head>
|   <body>
|     <b>
|     <p>
|       <b>
|       "TEST"

#data
<p id=a><b><p id=b></b>TEST
#errors
(1,8): expected-doctype-but-got-start-tag
(1,19): unexpected-end-tag
(1,23): adoption-agency-1.2
#document
| <html>
|   <head>
|   <body>
|     <p>
|       id="a"
|       <b>
|     <p>
|       id="b"
|       "TEST"

#data
<b id=a><p><b id=b></p></b>TEST
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): unexpected-end-tag
(1,27): adoption-agency-1.2
(1,31): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       id="a"
|       <p>
|         <b>
|           id="b"
|       "TEST"

#data
<!DOCTYPE html><title>U-test</title><body><div><p>Test<u></p></div></body>
#errors
(1,61): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "U-test"
|   <body>
|     <div>
|       <p>
|         "Test"
|         <u>

#data
<!DOCTYPE html><font><table></font></table></font>
#errors
(1,35): unexpected-end-tag-implies-table-voodoo
(1,35): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <font>
|       <table>

#data
<font><p>hello<b>cruel</font>world
#errors
(1,6): expected-doctype-but-got-start-tag
(1,29): adoption-agency-1.3
(1,29): adoption-agency-1.3
(1,34): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <font>
|     <p>
|       <font>
|         "hello"
|         <b>
|           "cruel"
|       <b>
|         "world"

#data
<b>Test</i>Test
#errors
(1,3): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag
(1,15): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "TestTest"

#data
<b>A<cite>B<div>C
#errors
(1,3): expected-doctype-but-got-start-tag
(1,17): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "A"
|       <cite>
|         "B"
|         <div>
|           "C"

#data
<b>A<cite>B<div>C</cite>D
#errors
(1,3): expected-doctype-but-got-start-tag
(1,24): unexpected-end-tag
(1,25): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "A"
|       <cite>
|         "B"
|         <div>
|           "CD"

#data
<b>A<cite>B<div>C</b>D
#errors
(1,3): expected-doctype-but-got-start-tag
(1,21): adoption-agency-1.3
(1,22): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "A"
|       <cite>
|         "B"
|     <div>
|       <b>
|         "C"
|       "D"

#data

#errors
(1,0): expected-doctype-but-got-eof
#document
| <html>
|   <head>
|   <body>

#data
<DIV>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,5): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>

#data
<DIV> abc
#errors
(1,5): expected-doctype-but-got-start-tag
(1,9): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc"

#data
<DIV> abc <B>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,13): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>

#data
<DIV> abc <B> def
#errors
(1,5): expected-doctype-but-got-start-tag
(1,17): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def"

#data
<DIV> abc <B> def <I>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,21): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>

#data
<DIV> abc <B> def <I> ghi
#errors
(1,5): expected-doctype-but-got-start-tag
(1,25): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi"

#data
<DIV> abc <B> def <I> ghi <P>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,29): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi "
|           <p>

#data
<DIV> abc <B> def <I> ghi <P> jkl
#errors
(1,5): expected-doctype-but-got-start-tag
(1,33): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi "
|           <p>
|             " jkl"

#data
<DIV> abc <B> def <I> ghi <P> jkl </B>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,38): adoption-agency-1.3
(1,38): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi "
|       <i>
|         <p>
|           <b>
|             " jkl "

#data
<DIV> abc <B> def <I> ghi <P> jkl </B> mno
#errors
(1,5): expected-doctype-but-got-start-tag
(1,38): adoption-agency-1.3
(1,42): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi "
|       <i>
|         <p>
|           <b>
|             " jkl "
|           " mno"

#data
<DIV> abc <B> def <I> ghi <P> jkl </B> mno </I>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,38): adoption-agency-1.3
(1,47): adoption-agency-1.3
(1,47): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi "
|       <i>
|       <p>
|         <i>
|           <b>
|             " jkl "
|           " mno "

#data
<DIV> abc <B> def <I> ghi <P> jkl </B> mno </I> pqr
#errors
(1,5): expected-doctype-but-got-start-tag
(1,38): adoption-agency-1.3
(1,47): adoption-agency-1.3
(1,51): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi "
|       <i>
|       <p>
|         <i>
|           <b>
|             " jkl "
|           " mno "
|         " pqr"

#data
<DIV> abc <B> def <I> ghi <P> jkl </B> mno </I> pqr </P>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,38): adoption-agency-1.3
(1,47): adoption-agency-1.3
(1,56): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi "
|       <i>
|       <p>
|         <i>
|           <b>
|             " jkl "
|           " mno "
|         " pqr "

#data
<DIV> abc <B> def <I> ghi <P> jkl </B> mno </I> pqr </P> stu
#errors
(1,5): expected-doctype-but-got-start-tag
(1,38): adoption-agency-1.3
(1,47): adoption-agency-1.3
(1,60): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " abc "
|       <b>
|         " def "
|         <i>
|           " ghi "
|       <i>
|       <p>
|         <i>
|           <b>
|             " jkl "
|           " mno "
|         " pqr "
|       " stu"

#data
<test attribute---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
#errors
(1,1040): expected-doctype-but-got-start-tag
(1,1040): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <test>
|       attribute----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------=""

#data
<a href="blah">aba<table><a href="foo">br<tr><td></td></tr>x</table>aoe
#errors
(1,15): expected-doctype-but-got-start-tag
(1,39): unexpected-start-tag-implies-table-voodoo
(1,39): unexpected-start-tag-implies-end-tag
(1,39): unexpected-end-tag
(1,45): foster-parenting-character-in-table
(1,45): foster-parenting-character-in-table
(1,68): foster-parenting-character-in-table
(1,71): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|       <a>
|         href="foo"
|         "br"
|       <a>
|         href="foo"
|         "x"
|       <table>
|         <tbody>
|           <tr>
|             <td>
|     <a>
|       href="foo"
|       "aoe"

#data
<a href="blah">aba<table><tr><td><a href="foo">br</td></tr>x</table>aoe
#errors
(1,15): expected-doctype-but-got-start-tag
(1,54): unexpected-cell-end-tag
(1,68): unexpected text in table
(1,71): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "abax"
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 href="foo"
|                 "br"
|       "aoe"

#data
<table><a href="blah">aba<tr><td><a href="foo">br</td></tr>x</table>aoe
#errors
(1,7): expected-doctype-but-got-start-tag
(1,22): unexpected-start-tag-implies-table-voodoo
(1,29): foster-parenting-character-in-table
(1,29): foster-parenting-character-in-table
(1,29): foster-parenting-character-in-table
(1,54): unexpected-cell-end-tag
(1,68): foster-parenting-character-in-table
(1,71): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|     <a>
|       href="blah"
|       "x"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <a>
|               href="foo"
|               "br"
|     <a>
|       href="blah"
|       "aoe"

#data
<a href=a>aa<marquee>aa<a href=b>bb</marquee>aa
#errors
(1,10): expected-doctype-but-got-start-tag
(1,45): end-tag-too-early
(1,47): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="a"
|       "aa"
|       <marquee>
|         "aa"
|         <a>
|           href="b"
|           "bb"
|       "aa"

#data
<wbr><strike><code></strike><code><strike></code>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,28): adoption-agency-1.3
(1,49): adoption-agency-1.3
(1,49): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <wbr>
|     <strike>
|       <code>
|     <code>
|       <code>
|         <strike>

#data
<!DOCTYPE html><spacer>foo
#errors
(1,26): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <spacer>
|       "foo"

#data
<title><meta></title><link><title><meta></title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <title>
|       "<meta>"
|     <link>
|     <title>
|       "<meta>"
|   <body>

#data
<style><!--</style><meta><script>--><link></script>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--"
|     <meta>
|     <script>
|       "--><link>"
|   <body>

#data
<body><body><base><link><meta><title><p></title><body><p></body>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,12): unexpected-start-tag
(1,54): unexpected-start-tag
#document
| <html>
|   <head>
|   <body>
|     <base>
|     <link>
|     <meta>
|     <title>
|       "<p>"
|     <p>

#data
<textarea><p></textarea>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<p>"

#data
<p><image></p>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,10): unexpected-start-tag-treated-as
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <img>

#data
<a><table><a></table><p><a><div><a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,13): unexpected-start-tag-implies-table-voodoo
(1,13): unexpected-start-tag-implies-end-tag
(1,13): adoption-agency-1.3
(1,27): unexpected-start-tag-implies-end-tag
(1,27): adoption-agency-1.2
(1,32): unexpected-end-tag
(1,35): unexpected-start-tag-implies-end-tag
(1,35): adoption-agency-1.2
(1,35): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|     <p>
|       <a>
|     <div>
|       <a>

#data
<head></p><meta><p>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,10): unexpected-end-tag
#document
| <html>
|   <head>
|     <meta>
|   <body>
|     <p>

#data
<head></html><meta><p>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,19): expected-eof-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <meta>
|     <p>

#data
<b><table><td><i></table>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,25): unexpected-cell-end-tag
(1,25): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <i>

#data
<b><table><td></b><i></table>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,18): unexpected-end-tag
(1,29): unexpected-cell-end-tag
(1,29): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <i>

#data
<h1><h2>
#errors
(1,4): expected-doctype-but-got-start-tag
(1,8): unexpected-start-tag
(1,8): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <h1>
|     <h2>

#data
<a><p><a></a></p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,9): unexpected-start-tag-implies-end-tag
(1,9): adoption-agency-1.3
(1,21): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <p>
|       <a>
|       <a>

#data
<b><button></b></button></b>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,15): adoption-agency-1.3
(1,28): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <b>
|     <button>
|       <b>

#data
<p><b><div><marquee></p></b></div>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag
(1,24): unexpected-end-tag
(1,28): unexpected-end-tag
(1,34): end-tag-too-early
(1,34): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <div>
|       <b>
|         <marquee>
|           <p>

#data
<script></script></div><title></title><p><p>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): unexpected-end-tag
#document
| <html>
|   <head>
|     <script>
|     <title>
|   <body>
|     <p>
|     <p>

#data
<p><hr></p>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <hr>
|     <p>

#data
<select><b><option><select><option></b></select>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): unexpected-start-tag-in-select
(1,27): unexpected-select-in-select
(1,39): unexpected-end-tag
(1,48): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|     <option>

#data
<html><head><title></title><body></body></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <title>
|   <body>

#data
<a><table><td><a><table></table><a></tr><a></table><a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,14): unexpected-cell-in-table-body
(1,35): unexpected-start-tag-implies-end-tag
(1,40): unexpected-cell-end-tag
(1,43): unexpected-start-tag-implies-table-voodoo
(1,43): unexpected-start-tag-implies-end-tag
(1,43): unexpected-end-tag
(1,54): unexpected-start-tag-implies-end-tag
(1,54): adoption-agency-1.2
(1,54): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|     <a>


#data
<ul><li><ul></li><li>a</li></ul></li></ul>
#errors
(1,4): expected-doctype-but-got-start-tag
(1,17): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <ul>
|           <li>
|             "a"

#data
<ul><li></li><div><li></div><li><li><div><li><address><li><b><em></b><li></ul>
#errors
(1,4): expected-doctype-but-got-start-tag
(1,45): end-tag-too-early
(1,58): end-tag-too-early
(1,69): adoption-agency-1.3
#document
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|       <div>
|         <li>
|       <li>
|       <li>
|         <div>
|       <li>
|         <address>
|       <li>
|         <b>
|           <em>
|       <li>

#data
<frameset><frame><frameset><frame></frameset><noframes></noframes></frameset>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <frameset>
|     <frame>
|     <frameset>
|       <frame>
|     <noframes>

#data
<h1><table><td><h3></table><h3></h1>
#errors
(1,4): expected-doctype-but-got-start-tag
(1,15): unexpected-cell-in-table-body
(1,27): unexpected-cell-end-tag
(1,31): unexpected-start-tag
(1,36): end-tag-too-early
#document
| <html>
|   <head>
|   <body>
|     <h1>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <h3>
|     <h3>

#data
<table><colgroup><col><colgroup><col><col><col><colgroup><col><col><thead><tr><td></table>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         <col>
|       <colgroup>
|         <col>
|         <col>
|         <col>
|       <colgroup>
|         <col>
|         <col>
|       <thead>
|         <tr>
|           <td>

#data
<table><col><tbody><col><tr><col><td><col></table><col>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): unexpected-cell-in-table-body
(1,55): unexpected-start-tag-ignored
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         <col>
|       <tbody>
|       <colgroup>
|         <col>
|       <tbody>
|         <tr>
|       <colgroup>
|         <col>
|       <tbody>
|         <tr>
|           <td>
|       <colgroup>
|         <col>

#data
<table><colgroup><tbody><colgroup><tr><colgroup><td><colgroup></table><colgroup>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,52): unexpected-cell-in-table-body
(1,80): unexpected-start-tag-ignored
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|       <tbody>
|       <colgroup>
|       <tbody>
|         <tr>
|       <colgroup>
|       <tbody>
|         <tr>
|           <td>
|       <colgroup>

#data
</strong></b></em></i></u></strike></s></blink></tt></pre></big></small></font></select></h1></h2></h3></h4></h5></h6></body></br></a></img></title></span></style></script></table></th></td></tr></frame></area></link></param></hr></input></col></base></meta></basefont></bgsound></embed></spacer></p></dd></dt></caption></colgroup></tbody></tfoot></thead></address></blockquote></center></dir></div></dl></fieldset></listing></menu></ol></ul></li></nobr></wbr></form></button></marquee></object></html></frameset></head></iframe></image></isindex></noembed></noframes></noscript></optgroup></option></plaintext></textarea>
#errors
(1,9): expected-doctype-but-got-end-tag
(1,9): unexpected-end-tag-before-html
(1,13): unexpected-end-tag-before-html
(1,18): unexpected-end-tag-before-html
(1,22): unexpected-end-tag-before-html
(1,26): unexpected-end-tag-before-html
(1,35): unexpected-end-tag-before-html
(1,39): unexpected-end-tag-before-html
(1,47): unexpected-end-tag-before-html
(1,52): unexpected-end-tag-before-html
(1,58): unexpected-end-tag-before-html
(1,64): unexpected-end-tag-before-html
(1,72): unexpected-end-tag-before-html
(1,79): unexpected-end-tag-before-html
(1,88): unexpected-end-tag-before-html
(1,93): unexpected-end-tag-before-html
(1,98): unexpected-end-tag-before-html
(1,103): unexpected-end-tag-before-html
(1,108): unexpected-end-tag-before-html
(1,113): unexpected-end-tag-before-html
(1,118): unexpected-end-tag-before-html
(1,130): unexpected-end-tag-after-body
(1,130): unexpected-end-tag-treated-as
(1,134): unexpected-end-tag
(1,140): unexpected-end-tag
(1,148): unexpected-end-tag
(1,155): unexpected-end-tag
(1,163): unexpected-end-tag
(1,172): unexpected-end-tag
(1,180): unexpected-end-tag
(1,185): unexpected-end-tag
(1,190): unexpected-end-tag
(1,195): unexpected-end-tag
(1,203): unexpected-end-tag
(1,210): unexpected-end-tag
(1,217): unexpected-end-tag
(1,225): unexpected-end-tag
(1,230): unexpected-end-tag
(1,238): unexpected-end-tag
(1,244): unexpected-end-tag
(1,251): unexpected-end-tag
(1,258): unexpected-end-tag
(1,269): unexpected-end-tag
(1,279): unexpected-end-tag
(1,287): unexpected-end-tag
(1,296): unexpected-end-tag
(1,300): unexpected-end-tag
(1,305): unexpected-end-tag
(1,310): unexpected-end-tag
(1,320): unexpected-end-tag
(1,331): unexpected-end-tag
(1,339): unexpected-end-tag
(1,347): unexpected-end-tag
(1,355): unexpected-end-tag
(1,365): end-tag-too-early
(1,378): end-tag-too-early
(1,387): end-tag-too-early
(1,393): end-tag-too-early
(1,399): end-tag-too-early
(1,404): end-tag-too-early
(1,415): end-tag-too-early
(1,425): end-tag-too-early
(1,432): end-tag-too-early
(1,437): end-tag-too-early
(1,442): end-tag-too-early
(1,447): unexpected-end-tag
(1,454): unexpected-end-tag
(1,460): unexpected-end-tag
(1,467): unexpected-end-tag
(1,476): end-tag-too-early
(1,486): end-tag-too-early
(1,495): end-tag-too-early
(1,513): expected-eof-but-got-end-tag
(1,513): unexpected-end-tag
(1,520): unexpected-end-tag
(1,529): unexpected-end-tag
(1,537): unexpected-end-tag
(1,547): unexpected-end-tag
(1,557): unexpected-end-tag
(1,568): unexpected-end-tag
(1,579): unexpected-end-tag
(1,590): unexpected-end-tag
(1,599): unexpected-end-tag
(1,611): unexpected-end-tag
(1,622): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <br>
|     <p>

#data
<table><tr></strong></b></em></i></u></strike></s></blink></tt></pre></big></small></font></select></h1></h2></h3></h4></h5></h6></body></br></a></img></title></span></style></script></table></th></td></tr></frame></area></link></param></hr></input></col></base></meta></basefont></bgsound></embed></spacer></p></dd></dt></caption></colgroup></tbody></tfoot></thead></address></blockquote></center></dir></div></dl></fieldset></listing></menu></ol></ul></li></nobr></wbr></form></button></marquee></object></html></frameset></head></iframe></image></isindex></noembed></noframes></noscript></optgroup></option></plaintext></textarea>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,20): unexpected-end-tag-implies-table-voodoo
(1,20): unexpected-end-tag
(1,24): unexpected-end-tag-implies-table-voodoo
(1,24): unexpected-end-tag
(1,29): unexpected-end-tag-implies-table-voodoo
(1,29): unexpected-end-tag
(1,33): unexpected-end-tag-implies-table-voodoo
(1,33): unexpected-end-tag
(1,37): unexpected-end-tag-implies-table-voodoo
(1,37): unexpected-end-tag
(1,46): unexpected-end-tag-implies-table-voodoo
(1,46): unexpected-end-tag
(1,50): unexpected-end-tag-implies-table-voodoo
(1,50): unexpected-end-tag
(1,58): unexpected-end-tag-implies-table-voodoo
(1,58): unexpected-end-tag
(1,63): unexpected-end-tag-implies-table-voodoo
(1,63): unexpected-end-tag
(1,69): unexpected-end-tag-implies-table-voodoo
(1,69): end-tag-too-early
(1,75): unexpected-end-tag-implies-table-voodoo
(1,75): unexpected-end-tag
(1,83): unexpected-end-tag-implies-table-voodoo
(1,83): unexpected-end-tag
(1,90): unexpected-end-tag-implies-table-voodoo
(1,90): unexpected-end-tag
(1,99): unexpected-end-tag-implies-table-voodoo
(1,99): unexpected-end-tag
(1,104): unexpected-end-tag-implies-table-voodoo
(1,104): end-tag-too-early
(1,109): unexpected-end-tag-implies-table-voodoo
(1,109): end-tag-too-early
(1,114): unexpected-end-tag-implies-table-voodoo
(1,114): end-tag-too-early
(1,119): unexpected-end-tag-implies-table-voodoo
(1,119): end-tag-too-early
(1,124): unexpected-end-tag-implies-table-voodoo
(1,124): end-tag-too-early
(1,129): unexpected-end-tag-implies-table-voodoo
(1,129): end-tag-too-early
(1,136): unexpected-end-tag-in-table-row
(1,141): unexpected-end-tag-implies-table-voodoo
(1,141): unexpected-end-tag-treated-as
(1,145): unexpected-end-tag-implies-table-voodoo
(1,145): unexpected-end-tag
(1,151): unexpected-end-tag-implies-table-voodoo
(1,151): unexpected-end-tag
(1,159): unexpected-end-tag-implies-table-voodoo
(1,159): unexpected-end-tag
(1,166): unexpected-end-tag-implies-table-voodoo
(1,166): unexpected-end-tag
(1,174): unexpected-end-tag-implies-table-voodoo
(1,174): unexpected-end-tag
(1,183): unexpected-end-tag-implies-table-voodoo
(1,183): unexpected-end-tag
(1,196): unexpected-end-tag
(1,201): unexpected-end-tag
(1,206): unexpected-end-tag
(1,214): unexpected-end-tag
(1,221): unexpected-end-tag
(1,228): unexpected-end-tag
(1,236): unexpected-end-tag
(1,241): unexpected-end-tag
(1,249): unexpected-end-tag
(1,255): unexpected-end-tag
(1,262): unexpected-end-tag
(1,269): unexpected-end-tag
(1,280): unexpected-end-tag
(1,290): unexpected-end-tag
(1,298): unexpected-end-tag
(1,307): unexpected-end-tag
(1,311): unexpected-end-tag
(1,316): unexpected-end-tag
(1,321): unexpected-end-tag
(1,331): unexpected-end-tag
(1,342): unexpected-end-tag
(1,350): unexpected-end-tag
(1,358): unexpected-end-tag
(1,366): unexpected-end-tag
(1,376): end-tag-too-early
(1,389): end-tag-too-early
(1,398): end-tag-too-early
(1,404): end-tag-too-early
(1,410): end-tag-too-early
(1,415): end-tag-too-early
(1,426): end-tag-too-early
(1,436): end-tag-too-early
(1,443): end-tag-too-early
(1,448): end-tag-too-early
(1,453): end-tag-too-early
(1,458): unexpected-end-tag
(1,465): unexpected-end-tag
(1,471): unexpected-end-tag
(1,478): unexpected-end-tag
(1,487): end-tag-too-early
(1,497): end-tag-too-early
(1,506): end-tag-too-early
(1,524): expected-eof-but-got-end-tag
(1,524): unexpected-end-tag
(1,531): unexpected-end-tag
(1,540): unexpected-end-tag
(1,548): unexpected-end-tag
(1,558): unexpected-end-tag
(1,568): unexpected-end-tag
(1,579): unexpected-end-tag
(1,590): unexpected-end-tag
(1,601): unexpected-end-tag
(1,610): unexpected-end-tag
(1,622): unexpected-end-tag
(1,633): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <br>
|     <table>
|       <tbody>
|         <tr>
|     <p>

#data
<frameset>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,10): eof-in-frameset
#document
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html>Test
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "Test"

#data
<textarea>test</div>test
#errors
(1,10): expected-doctype-but-got-start-tag
(1,24): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "test</div>test"

#data
<table><td>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,11): unexpected-cell-in-table-body
(1,11): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<table><td>test</tbody></table>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,11): unexpected-cell-in-table-body
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "test"

#data
<frame>test
#errors
(1,7): expected-doctype-but-got-start-tag
(1,7): unexpected-start-tag-ignored
#document
| <html>
|   <head>
|   <body>
|     "test"

#data
<!DOCTYPE html><frameset>test
#errors
(1,29): unexpected-char-in-frameset
(1,29): unexpected-char-in-frameset
(1,29): unexpected-char-in-frameset
(1,29): unexpected-char-in-frameset
(1,29): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><frameset> te st
#errors
(1,29): unexpected-char-in-frameset
(1,29): unexpected-char-in-frameset
(1,29): unexpected-char-in-frameset
(1,29): unexpected-char-in-frameset
(1,29): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|     "  "

#data
<!DOCTYPE html><frameset></frameset> te st
#errors
(1,29): unexpected-char-after-frameset
(1,29): unexpected-char-after-frameset
(1,29): unexpected-char-after-frameset
(1,29): unexpected-char-after-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   "  "

#data
<!DOCTYPE html><frameset><!DOCTYPE html>
#errors
(1,40): unexpected-doctype
(1,40): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><font><p><b>test</font>
#errors
(1,38): adoption-agency-1.3
(1,38): adoption-agency-1.3
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <font>
|     <p>
|       <font>
|         <b>
|           "test"

#data
<!DOCTYPE html><dt><div><dd>
#errors
(1,28): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <dt>
|       <div>
|     <dd>

#data
<script></x
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</x"
|   <body>

#data
<table><plaintext><td>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,18): unexpected-start-tag-implies-table-voodoo
(1,22): foster-parenting-character-in-table
(1,22): foster-parenting-character-in-table
(1,22): foster-parenting-character-in-table
(1,22): foster-parenting-character-in-table
(1,22): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "<td>"
|     <table>

#data
<plaintext></plaintext>
#errors
(1,11): expected-doctype-but-got-start-tag
(1,23): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "</plaintext>"

#data
<!DOCTYPE html><table><tr>TEST
#errors
(1,30): foster-parenting-character-in-table
(1,30): foster-parenting-character-in-table
(1,30): foster-parenting-character-in-table
(1,30): foster-parenting-character-in-table
(1,30): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "TEST"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE html><body t1=1><body t2=2><body t3=3 t4=4>
#errors
(1,37): unexpected-start-tag
(1,53): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     t1="1"
|     t2="2"
|     t3="3"
|     t4="4"

#data
</b test
#errors
(1,8): eof-in-attribute-name
(1,8): expected-doctype-but-got-eof
#new-errors
(1:9) eof-in-tag
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html></b test<b =amp>X
#errors
(1,24): invalid-character-in-attribute-name
(1,32): named-entity-without-semicolon
(1,33): attributes-in-end-tag
(1,33): unexpected-end-tag-before-html
#new-errors
(1:24) unexpected-character-in-attribute-name
(1:33) missing-semicolon-after-character-reference
(1:33) end-tag-with-attributes
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"

#data
<!doctypehtml><scrIPt type=text/x-foobar;baz>X</SCRipt
#errors
(1,9): need-space-after-doctype
(1,54): expected-named-closing-tag-but-got-eof
#new-errors
(1:10) missing-whitespace-before-doctype-name
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       type="text/x-foobar;baz"
|       "X</SCRipt"
|   <body>

#data
#
#errors
(1,2): expected-numeric-entity
(1,2): expected-doctype-but-got-chars
#new-errors
(1:3) absence-of-digits-in-numeric-character-reference
#document
| <html>
|   <head>
|   <body>
|     "#"

#data
<!DOCTYPE html><table><tr>TEST
#errors
(1,30): foster-parenting-character-in-table
(1,30): foster-parenting-character-in-table
(1,30): foster-parenting-character-in-table
(1,30): foster-parenting-character-in-table
(1,30): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "TEST"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE html><body t1=1><body t2=2><body t3=3 t4=4>
#errors
(1,37): unexpected-start-tag
(1,53): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     t1="1"
|     t2="2"
|     t3="3"
|     t4="4"

#data
</b test
#errors
(1,8): eof-in-attribute-name
(1,8): expected-doctype-but-got-eof
#new-errors
(1:9) eof-in-tag
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html></b test<b =amp>X
#errors
(1,24): invalid-character-in-attribute-name
(1,32): named-entity-without-semicolon
(1,33): attributes-in-end-tag
(1,33): unexpected-end-tag-before-html
#new-errors
(1:24) unexpected-character-in-attribute-name
(1:33) missing-semicolon-after-character-reference
(1:33) end-tag-with-attributes
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"

#data
<!doctypehtml><scrIPt type=text/x-foobar;baz>X</SCRipt
#errors
(1,9): need-space-after-doctype
(1,54): expected-named-closing-tag-but-got-eof
#new-errors
(1:10) missing-whitespace-before-doctype-name
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       type="text/x-foobar;baz"
|       "X</SCRipt"
|   <body>



#data
#
#errors
(1,2): expected-numeric-entity
(1,2): expected-doctype-but-got-chars
#new-errors
(1:3) absence-of-digits-in-numeric-character-reference
#document
| <html>
|   <head>
|   <body>
|     "#"

#data
#X
#errors
(1,3): expected-numeric-entity
(1,3): expected-doctype-but-got-chars
#new-errors
(1:4) absence-of-digits-in-numeric-character-reference
#document
| <html>
|   <head>
|   <body>
|     "#X"

#data
x-test
#errors
(1,2): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "x-test"

#data
<!doctypehtml><p><li>
#errors
(1,9): need-space-after-doctype
#new-errors
(1:10) missing-whitespace-before-doctype-name
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <li>

#data
<!doctypehtml><p><dt>
#errors
(1,9): need-space-after-doctype
#new-errors
(1:10) missing-whitespace-before-doctype-name
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <dt>

#data
&#45
#errors
(1,4): numeric-entity-without-semicolon
(1,4): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "-"

#data
&x-test
#errors
(1,2): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "&x-test"

#data
<!doctypehtml><p><li>
#errors
(1,9): need-space-after-doctype
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <li>

#data
<!doctypehtml><p><dt>
#errors
(1,9): need-space-after-doctype
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <dt>

#data
<!doctypehtml><p><dd>
#errors
(1,9): need-space-after-doctype
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <dd>

#data
<!doctypehtml><p><form>
#errors
(1,9): need-space-after-doctype
(1,23): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <form>

#data
<!DOCTYPE html><p></P>X
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     "X"

#data
&AMP
#errors
(1,4): named-entity-without-semicolon
(1,4): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "&"

#data
&AMp;
#errors
(1,3): expected-named-entity
(1,3): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "&AMp;"

#data
<!DOCTYPE html><html><head></head><body><thisISasillyTESTelementNameToMakeSureCrazyTagNamesArePARSEDcorrectLY>
#errors
(1,110): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <thisisasillytestelementnametomakesurecrazytagnamesareparsedcorrectly>

#data
<!DOCTYPE html>X</body>X
#errors
(1,24): unexpected-char-after-body
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "XX"

#data
<!DOCTYPE html><!-- X
#errors
(1,21): eof-in-comment
#document
| <!DOCTYPE html>
| <!--  X -->
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><table><caption>test TEST</caption><td>test
#errors
(1,54): unexpected-cell-in-table-body
(1,58): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "test TEST"
|       <tbody>
|         <tr>
|           <td>
|             "test"

#data
<!DOCTYPE html><select><option><optgroup>
#errors
(1,41): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <optgroup>

#data
<!DOCTYPE html><select><optgroup><option></optgroup><option><select><option>
#errors
(1,68): unexpected-select-in-select
(1,76): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|         <option>
|       <option>
|     <option>

#data
<!DOCTYPE html><select><optgroup><option><optgroup>
#errors
(1,51): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|         <option>
|       <optgroup>

#data
<!DOCTYPE html><datalist><option>foo</datalist>bar
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <datalist>
|       <option>
|         "foo"
|     "bar"

#data
<!DOCTYPE html><font><input><input></font>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <font>
|       <input>
|       <input>

#data
<!DOCTYPE html><!-- XXX - XXX -->
#errors
#document
| <!DOCTYPE html>
| <!--  XXX - XXX  -->
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><!-- XXX - XXX
#errors
(1,29): eof-in-comment
#document
| <!DOCTYPE html>
| <!--  XXX - XXX -->
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><!-- XXX - XXX - XXX -->
#errors
#document
| <!DOCTYPE html>
| <!--  XXX - XXX - XXX  -->
| <html>
|   <head>
|   <body>

#data
test
test
#errors
(2,4): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "test
test"

#data
<!DOCTYPE html><body><title>test</body></title>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <title>
|       "test</body>"

#data
<!DOCTYPE html><body><title>X</title><meta name=z><link rel=foo><style>
x { content:"</style" } </style>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <title>
|       "X"
|     <meta>
|       name="z"
|     <link>
|       rel="foo"
|     <style>
|       "
x { content:"</style" } "

#data
<!DOCTYPE html><select><optgroup></optgroup></select>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>

#data
 
 
#errors
(2,1): expected-doctype-but-got-eof
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html>  <html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><script>
</script>  <title>x</title>  </head>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "
"
|     "  "
|     <title>
|       "x"
|     "  "
|   <body>

#data
<!DOCTYPE html><html><body><html id=x>
#errors
(1,38): non-html-root
#document
| <!DOCTYPE html>
| <html>
|   id="x"
|   <head>
|   <body>

#data
<!DOCTYPE html>X</body><html id="x">
#errors
(1,36): non-html-root
#document
| <!DOCTYPE html>
| <html>
|   id="x"
|   <head>
|   <body>
|     "X"

#data
<!DOCTYPE html><head><html id=x>
#errors
(1,32): non-html-root
#document
| <!DOCTYPE html>
| <html>
|   id="x"
|   <head>
|   <body>

#data
<!DOCTYPE html>X</html>X
#errors
(1,24): expected-eof-but-got-char
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "XX"

#data
<!DOCTYPE html>X</html> 
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X "

#data
<!DOCTYPE html>X</html><p>X
#errors
(1,26): expected-eof-but-got-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"
|     <p>
|       "X"

#data
<!DOCTYPE html>X<p/x/y/z>
#errors
(1,19): unexpected-character-after-solidus-in-tag
(1,21): unexpected-character-after-solidus-in-tag
(1,23): unexpected-character-after-solidus-in-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"
|     <p>
|       x=""
|       y=""
|       z=""

#data
<!DOCTYPE html><!--x--
#errors
(1,22): eof-in-comment-double-dash
#document
| <!DOCTYPE html>
| <!-- x -->
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><table><tr><td></p></table>
#errors
(1,34): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <p>

#data
<!DOCTYPE <!DOCTYPE HTML>><!--<!--x-->-->
#errors
(1,20): expected-space-or-right-bracket-in-doctype
(1,25): unknown-doctype
(1,35): unexpected-char-in-comment
#document
| <!DOCTYPE <!doctype>
| <html>
|   <head>
|   <body>
|     ">"
|     <!-- <!--x -->
|     "-->"

#data
<!doctype html><div><form></form><div></div></div>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>
|       <form>
|       <div>

#data
<head></head><style></style>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,20): unexpected-start-tag-out-of-my-head
#document
| <html>
|   <head>
|     <style>
|   <body>

#data
<head></head><script></script>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,21): unexpected-start-tag-out-of-my-head
#document
| <html>
|   <head>
|     <script>
|   <body>

#data
<head></head><!-- --><style></style><!-- --><script></script>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,28): unexpected-start-tag-out-of-my-head
(1,52): unexpected-start-tag-out-of-my-head
#document
| <html>
|   <head>
|     <style>
|     <script>
|   <!--   -->
|   <!--   -->
|   <body>

#data
<head></head><!-- -->x<style></style><!-- --><script></script>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <!--   -->
|   <body>
|     "x"
|     <style>
|     <!--   -->
|     <script>

#data
<!DOCTYPE html><html><head></head><body><pre>
</pre></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>

#data
<!DOCTYPE html><html><head></head><body><pre>
foo</pre></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>
|       "foo"

#data
<!DOCTYPE html><html><head></head><body><pre>

foo</pre></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>
|       "
foo"

#data
<!DOCTYPE html><html><head></head><body><pre>
foo
</pre></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>
|       "foo
"

#data
<!DOCTYPE html><html><head></head><body><pre>x</pre><span>
</span></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>
|       "x"
|     <span>
|       "
"
#data
<!DOCTYPE html><html><head></head><body><pre>x
y</pre></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>
|       "x
y"

#data
<!DOCTYPE html><html><head></head><body><pre>x<div>
y</pre></body></html>
#errors
(2,7): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>
|       "x"
|       <div>
|         "
y"

#data
<!DOCTYPE html><pre>&#x0a;&#x0a;A</pre>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>
|       "
A"

#data
<!DOCTYPE html><HTML><META><HEAD></HEAD></HTML>
#errors
(1,33): two-heads-are-not-better-than-one
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <meta>
|   <body>

#data
<!DOCTYPE html><HTML><HEAD><head></HEAD></HTML>
#errors
(1,33): two-heads-are-not-better-than-one
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<textarea>foo<span>bar</span><i>baz
#errors
(1,10): expected-doctype-but-got-start-tag
(1,35): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "foo<span>bar</span><i>baz"

#data
<title>foo<span>bar</em><i>baz
#errors
(1,7): expected-doctype-but-got-start-tag
(1,30): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <title>
|       "foo<span>bar</em><i>baz"
|   <body>

#data
<!DOCTYPE html><textarea>
</textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>

#data
<!DOCTYPE html><textarea>
foo</textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "foo"

#data
<!DOCTYPE html><textarea>

foo</textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "
foo"

#data
<!DOCTYPE html><html><head></head><body><ul><li><div><p><li></ul></body></html>
#errors
(1,60): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <div>
|           <p>
|       <li>

#data
<!doctype html><nobr><nobr><nobr>
#errors
(1,27): unexpected-start-tag-implies-end-tag
(1,33): unexpected-start-tag-implies-end-tag
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <nobr>
|     <nobr>
|     <nobr>

#data
<!doctype html><nobr><nobr></nobr><nobr>
#errors
(1,27): unexpected-start-tag-implies-end-tag
(1,40): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <nobr>
|     <nobr>
|     <nobr>


#data
<!doctype html><html><body><p><table></table></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <table>

#data
<p><table></table>
#errors
(1,3): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <table>

#data
direct div content
#errors
#document-fragment
div
#document
| "direct div content"

#data
direct textarea content
#errors
#document-fragment
textarea
#document
| "direct textarea content"

#data
textarea content with <em>pseudo</em> <foo>markup
#errors
#document-fragment
textarea
#document
| "textarea content with <em>pseudo</em> <foo>markup"

#data
this is &#x0043;DATA inside a <style> element
#errors
#document-fragment
style
#document
| "this is &#x0043;DATA inside a <style> element"

#data
</plaintext>
#errors
#document-fragment
plaintext
#document
| "</plaintext>"

#data
setting html's innerHTML
#errors
#document-fragment
html
#document
| <head>
| <body>
|   "setting html's innerHTML"

#data
<title>setting head's innerHTML</title>
#errors
#document-fragment
head
#document
| <title>
|   "setting head's innerHTML"

#data
<style> <!-- </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       " <!-- "
|   <body>
|     "x"

#data
<style> <!-- </style> --> </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
(1,34): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       " <!-- "
|     " "
|   <body>
|     "--> x"

#data
<style> <!--> </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       " <!--> "
|   <body>
|     "x"

#data
<style> <!---> </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       " <!---> "
|   <body>
|     "x"

#data
<iframe> <!---> </iframe>x
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <iframe>
|       " <!---> "
|     "x"

#data
<iframe> <!--- </iframe>->x</iframe> --> </iframe>x
#errors
(1,8): expected-doctype-but-got-start-tag
(1,36): unexpected-end-tag
(1,50): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <iframe>
|       " <!--- "
|     "->x --> x"

#data
<script> <!-- </script> --> </script>x
#errors
(1,8): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag
#document
| <html>
|   <head>
|     <script>
|       " <!-- "
|     " "
|   <body>
|     "--> x"

#data
<title> <!-- </title> --> </title>x
#errors
(1,7): expected-doctype-but-got-start-tag
(1,34): unexpected-end-tag
#document
| <html>
|   <head>
|     <title>
|       " <!-- "
|     " "
|   <body>
|     "--> x"

#data
<textarea> <!--- </textarea>->x</textarea> --> </textarea>x
#errors
(1,10): expected-doctype-but-got-start-tag
(1,42): unexpected-end-tag
(1,58): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       " <!--- "
|     "->x --> x"

#data
<style> <!</-- </style>x
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       " <!</-- "
|   <body>
|     "x"

#data
<p><xmp></xmp>
#errors
(1,3): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <p>
|     <xmp>

#data
<xmp> <!-- > --> </xmp>
#errors
(1,5): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <xmp>
|       " <!-- > --> "

#data
<title>&amp;</title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <title>
|       "&"
|   <body>

#data
<title><!--&amp;--></title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <title>
|       "<!--&-->"
|   <body>

#data
<title><!--</title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <title>
|       "<!--"
|   <body>

#data
<noscript><!--</noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,39): unexpected-end-tag
#script-on
#document
| <html>
|   <head>
|     <noscript>
|       "<!--"
|   <body>
|     "-->"

#data
<noscript><!--</noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
#script-off
#document
| <html>
|   <head>
|     <noscript>
|       <!-- </noscript> -->
|   <body>

#data
<!doctype html></head> <head>
#errors
(1,29): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   " "
|   <body>

#data
<!doctype html><form><div></form><div>
#errors
(1,33): end-tag-too-early-ignored
(1,38): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         <div>

#data
<!doctype html><title>&amp;</title>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "&"
|   <body>

#data
<!doctype html><title><!--&amp;--></title>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "<!--&-->"
|   <body>

#data
<!doctype>
#errors
(1,9): need-space-after-doctype
(1,10): expected-doctype-name-but-got-right-bracket
(1,10): unknown-doctype
#document
| <!DOCTYPE >
| <html>
|   <head>
|   <body>

#data
<!---x
#errors
(1,6): eof-in-comment
(1,6): expected-doctype-but-got-eof
#document
| <!-- -x -->
| <html>
|   <head>
|   <body>

#data
<body>
<div>
#errors
(1,6): unexpected-start-tag
(2,5): expected-closing-tag-but-got-eof
#document-fragment
div
#document
| "
"
| <div>

#data
<frameset></frameset>
foo
#errors
(1,10): expected-doctype-but-got-start-tag
(2,1): unexpected-char-after-frameset
(2,2): unexpected-char-after-frameset
(2,3): unexpected-char-after-frameset
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
<noframes>
#errors
(1,10): expected-doctype-but-got-start-tag
(2,10): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <frameset>
|   "
"
|   <noframes>

#data
<frameset></frameset>
<div>
#errors
(1,10): expected-doctype-but-got-start-tag
(2,5): unexpected-start-tag-after-frameset
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
</html>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
</div>
#errors
(1,10): expected-doctype-but-got-start-tag
(2,6): unexpected-end-tag-after-frameset
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<form><form>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,12): unexpected-start-tag
(1,12): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <form>

#data
<button><button>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,16): unexpected-start-tag-implies-end-tag
(1,16): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <button>
|     <button>

#data
<table><tr><td></th>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,20): unexpected-end-tag
(1,20): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<table><caption><td>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,20): unexpected-cell-in-table-body
(1,20): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|       <tbody>
|         <tr>
|           <td>

#data
<table><caption><div>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,21): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <div>

#data
</caption><div>
#errors
(1,10): XXX-undefined-error
(1,15): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <div>

#data
<table><caption><div></caption>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,31): expected-one-end-tag-but-got-another
(1,31): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <div>

#data
<table><caption></table>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>

#data
</table><div>
#errors
(1,8): unexpected-end-tag
(1,13): expected-closing-tag-but-got-eof
#document-fragment
caption
#document
| <div>

#data
<table><caption></body></col></colgroup></html></tbody></td></tfoot></th></thead></tr>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,23): unexpected-end-tag
(1,29): unexpected-end-tag
(1,40): unexpected-end-tag
(1,47): unexpected-end-tag
(1,55): unexpected-end-tag
(1,60): unexpected-end-tag
(1,68): unexpected-end-tag
(1,73): unexpected-end-tag
(1,81): unexpected-end-tag
(1,86): unexpected-end-tag
(1,86): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>

#data
<table><caption><div></div>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,27): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <div>

#data
<table><tr><td></body></caption></col></colgroup></html>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,22): unexpected-end-tag
(1,32): unexpected-end-tag
(1,38): unexpected-end-tag
(1,49): unexpected-end-tag
(1,56): unexpected-end-tag
(1,56): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
</table></tbody></tfoot></thead></tr><div>
#errors
(1,8): unexpected-end-tag
(1,16): unexpected-end-tag
(1,24): unexpected-end-tag
(1,32): unexpected-end-tag
(1,37): unexpected-end-tag
(1,42): expected-closing-tag-but-got-eof
#document-fragment
td
#document
| <div>

#data
<table><colgroup>foo
#errors
(1,7): expected-doctype-but-got-start-tag
(1,18): foster-parenting-character-in-table
(1,19): foster-parenting-character-in-table
(1,20): foster-parenting-character-in-table
(1,20): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     "foo"
|     <table>
|       <colgroup>

#data
foo<col>
#errors
(1,1): unexpected-character-in-colgroup
(1,2): unexpected-character-in-colgroup
(1,3): unexpected-character-in-colgroup
#document-fragment
colgroup
#document
| <col>

#data
<table><colgroup></col>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,23): no-end-tag
(1,23): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>

#data
<frameset><div>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,15): unexpected-start-tag-in-frameset
(1,15): eof-in-frameset
#document
| <html>
|   <head>
|   <frameset>

#data
</frameset><frame>
#errors
(1,11): unexpected-frameset-in-frameset-innerhtml
#document-fragment
frameset
#document
| <frame>

#data
<frameset></div>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,16): unexpected-end-tag-in-frameset
(1,16): eof-in-frameset
#document
| <html>
|   <head>
|   <frameset>

#data
</body><div>
#errors
(1,7): unexpected-close-tag
(1,12): expected-closing-tag-but-got-eof
#document-fragment
body
#document
| <div>

#data
<table><tr><div>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,16): unexpected-start-tag-implies-table-voodoo
(1,16): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <div>
|     <table>
|       <tbody>
|         <tr>

#data
</tr><td>
#errors
(1,5): unexpected-end-tag
#document-fragment
tr
#document
| <td>

#data
</tbody></tfoot></thead><td>
#errors
(1,8): unexpected-end-tag
(1,16): unexpected-end-tag
(1,24): unexpected-end-tag
#document-fragment
tr
#document
| <td>

#data
<table><tr><div><td>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,16): foster-parenting-start-tag
(1,20): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<caption><col><colgroup><tbody><tfoot><thead><tr>
#errors
(1,9): unexpected-start-tag
(1,14): unexpected-start-tag
(1,24): unexpected-start-tag
(1,31): unexpected-start-tag
(1,38): unexpected-start-tag
(1,45): unexpected-start-tag
#document-fragment
tbody
#document
| <tr>

#data
<table><tbody></thead>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,22): unexpected-end-tag-in-table-body
(1,22): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>

#data
</table><tr>
#errors
(1,8): unexpected-end-tag
#document-fragment
tbody
#document
| <tr>

#data
<table><tbody></body></caption></col></colgroup></html></td></th></tr>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,21): unexpected-end-tag-in-table-body
(1,31): unexpected-end-tag-in-table-body
(1,37): unexpected-end-tag-in-table-body
(1,48): unexpected-end-tag-in-table-body
(1,55): unexpected-end-tag-in-table-body
(1,60): unexpected-end-tag-in-table-body
(1,65): unexpected-end-tag-in-table-body
(1,70): unexpected-end-tag-in-table-body
(1,70): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>

#data
<table><tbody></div>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,20): unexpected-end-tag-implies-table-voodoo
(1,20): end-tag-too-early
(1,20): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>

#data
<table><table>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,14): unexpected-start-tag-implies-end-tag
(1,14): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <table>
|     <table>

#data
<table></body></caption></col></colgroup></html></tbody></td></tfoot></th></thead></tr>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,14): unexpected-end-tag
(1,24): unexpected-end-tag
(1,30): unexpected-end-tag
(1,41): unexpected-end-tag
(1,48): unexpected-end-tag
(1,56): unexpected-end-tag
(1,61): unexpected-end-tag
(1,69): unexpected-end-tag
(1,74): unexpected-end-tag
(1,82): unexpected-end-tag
(1,87): unexpected-end-tag
(1,87): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <table>

#data
</table><tr>
#errors
(1,8): unexpected-end-tag
#document-fragment
table
#document
| <tbody>
|   <tr>

#data
<body></body></html>
#errors
(1,20): unexpected-end-tag-after-body-innerhtml
#document-fragment
html
#document
| <head>
| <body>

#data
<html><frameset></frameset></html> 
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <frameset>
|   " "

#data
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"><html></html>
#errors
#document
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "">
| <html>
|   <head>
|   <body>

#data
<param><frameset></frameset>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,17): unexpected-start-tag
#document
| <html>
|   <head>
|   <frameset>

#data
</html><frameset></frameset>
#errors
(1,7): expected-doctype-but-got-end-tag
(1,17): expected-eof-but-got-start-tag
(1,17): unexpected-start-tag
#document
| <html>
|   <head>
|   <frameset>

#data
</body><frameset></frameset>
#errors
(1,7): expected-doctype-but-got-end-tag
(1,17): unexpected-start-tag-after-body
(1,17): unexpected-start-tag
#document
| <html>
|   <head>
|   <frameset>

#data
<!doctype html><body><title>X</title>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <title>
|       "X"

#data
<!doctype html><table><title>X</title></table>
#errors
(1,29): unexpected-start-tag-implies-table-voodoo
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <title>
|       "X"
|     <table>

#data
<!doctype html><head></head><title>X</title>
#errors
(1,35): unexpected-start-tag-out-of-my-head
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "X"
|   <body>

#data
<!doctype html></head><title>X</title>
#errors
(1,29): unexpected-start-tag-out-of-my-head
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "X"
|   <body>

#data
<!doctype html><table><meta></table>
#errors
(1,28): unexpected-start-tag-implies-table-voodoo
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <meta>
|     <table>

#data
<!doctype html><table>X<tr><td><table> <meta></table></table>
#errors
unexpected text in table
(1,45): unexpected-start-tag-implies-table-voodoo
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <meta>
|             <table>
|               " "

#data
<!doctype html><html> <head>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!doctype html> <head>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!doctype html><table><style> <tr>x </style> </table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <style>
|         " <tr>x "
|       " "

#data
<!doctype html><table><TBODY><script> <tr>x </script> </table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <script>
|           " <tr>x "
|         " "

#data
<!doctype html><p><applet><p>X</p></applet>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <p>
|           "X"

#data
<!doctype html><p><object type="application/x-non-existant-plugin"><p>X</p></object>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <object>
|         type="application/x-non-existant-plugin"
|         <p>
|           "X"

#data
<!doctype html><listing>
X</listing>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <listing>
|       "X"

#data
<!doctype html><select><input>X
#errors
(1,30): unexpected-input-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     <input>
|     "X"

#data
<!doctype html><select><select>X
#errors
(1,31): unexpected-select-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     "X"

#data
<!doctype html><table><input type=hidDEN></table>
#errors
(1,41): unexpected-hidden-input-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <input>
|         type="hidDEN"

#data
<!doctype html><table>X<input type=hidDEN></table>
#errors
(1,23): foster-parenting-character
(1,42): unexpected-hidden-input-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"
|     <table>
|       <input>
|         type="hidDEN"

#data
<!doctype html><table>  <input type=hidDEN></table>
#errors
(1,43): unexpected-hidden-input-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "  "
|       <input>
|         type="hidDEN"

#data
<!doctype html><table>  <input type='hidDEN'></table>
#errors
(1,45): unexpected-hidden-input-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "  "
|       <input>
|         type="hidDEN"

#data
<!doctype html><table><input type=" hidden"><input type=hidDEN></table>
#errors
(1,44): unexpected-start-tag-implies-table-voodoo
(1,63): unexpected-hidden-input-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <input>
|       type=" hidden"
|     <table>
|       <input>
|         type="hidDEN"

#data
<!doctype html><table><select>X<tr>
#errors
(1,30): unexpected-start-tag-implies-table-voodoo
(1,35): unexpected-table-element-start-tag-in-select-in-table
(1,35): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "X"
|     <table>
|       <tbody>
|         <tr>

#data
<!doctype html><select>X</select>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "X"

#data
<!DOCTYPE hTmL><html></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<body>X</body></body>
#errors
(1,21): unexpected-end-tag-after-body
#document-fragment
html
#document
| <head>
| <body>
|   "X"

#data
<div><p>a</x> b
#errors
(1,5): expected-doctype-but-got-start-tag
(1,13): unexpected-end-tag
(1,15): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <p>
|         "a b"

#data
<table><tr><td><code></code> </table>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <code>
|             " "

#data
<table><b><tr><td>aaa</td></tr>bbb</table>ccc
#errors
(1,7): expected-doctype-but-got-start-tag
(1,10): foster-parenting-start-tag
(1,32): foster-parenting-character
(1,33): foster-parenting-character
(1,34): foster-parenting-character
(1,45): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|     <b>
|       "bbb"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "aaa"
|     <b>
|       "ccc"

#data
A<table><tr> B</tr> B</table>
#errors
(1,1): expected-doctype-but-got-chars
(1,13): foster-parenting-character
(1,14): foster-parenting-character
(1,20): foster-parenting-character
(1,21): foster-parenting-character
#document
| <html>
|   <head>
|   <body>
|     "A B B"
|     <table>
|       <tbody>
|         <tr>

#data
A<table><tr> B</tr> </em>C</table>
#errors
(1,1): expected-doctype-but-got-chars
(1,13): foster-parenting-character
(1,14): foster-parenting-character
(1,20): foster-parenting-character
(1,25): unexpected-end-tag
(1,26): foster-parenting-character
#document
| <html>
|   <head>
|   <body>
|     "A BC"
|     <table>
|       <tbody>
|         <tr>
|         " "

#data
<select><keygen>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,16): unexpected-input-in-select
#document
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>

#data
<div>
<div></div>
</span>x
#errors
(1,5): expected-doctype-but-got-start-tag
(3,7): unexpected-end-tag
(3,8): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "
"
|       <div>
|       "
x"

#data
<div>x<div></div>
</span>x
#errors
(1,5): expected-doctype-but-got-start-tag
(2,7): unexpected-end-tag
(2,8): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "
x"

#data
<div>x<div></div>x</span>x
#errors
(1,5): expected-doctype-but-got-start-tag
(1,25): unexpected-end-tag
(1,26): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "xx"

#data
<div>x<div></div>y</span>z
#errors
(1,5): expected-doctype-but-got-start-tag
(1,25): unexpected-end-tag
(1,26): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "yz"

#data
<table><div>x<div></div>x</span>x
#errors
(1,7): expected-doctype-but-got-start-tag
(1,12): foster-parenting-start-tag
(1,13): foster-parenting-character
(1,18): foster-parenting-start-tag
(1,24): foster-parenting-end-tag
(1,25): foster-parenting-start-tag
(1,32): foster-parenting-end-tag
(1,32): unexpected-end-tag
(1,33): foster-parenting-character
(1,33): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "xx"
|     <table>

#data
x<table>x
#errors
(1,1): expected-doctype-but-got-chars
(1,9): foster-parenting-character
(1,9): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     "xx"
|     <table>

#data
x<table><table>x
#errors
(1,1): expected-doctype-but-got-chars
(1,15): unexpected-start-tag-implies-end-tag
(1,16): foster-parenting-character
(1,16): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|     "x"
|     <table>

#data
<b>a<div></div><div></b>y
#errors
(1,3): expected-doctype-but-got-start-tag
(1,24): adoption-agency-1.3
(1,25): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "a"
|       <div>
|     <div>
|       <b>
|       "y"

#data
<a><div><p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,15): adoption-agency-1.3
(1,15): adoption-agency-1.3
(1,15): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|       <p>
|         <a>

#data
<div>x<div></div>
</span>x
#errors
(1,5): expected-doctype-but-got-start-tag
(2,7): unexpected-end-tag
(2,8): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "
x"

#data
<div>x<div></div>x</span>x
#errors
(1,5): expected-doctype-but-got-start-tag
(1,25): unexpected-end-tag
(1,26): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "xx"

#data
<div>x<div></div>y</span>z
#errors
(1,5): expected-doctype-but-got-start-tag
(1,25): unexpected-end-tag
(1,26): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "yz"

#data
<table><div>x<div></div>x</span>x
#errors
(1,7): expected-doctype-but-got-start-tag
(1,12): foster-parenting-start-tag
(1,13): foster-parenting-character
(1,18): foster-parenting-start-tag
(1,24): foster-parenting-end-tag
(1,25): foster-parenting-start-tag
(1,32): foster-parenting-end-tag
(1,32): unexpected-end-tag
(1,33): foster-parenting-character
(1,33): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "xx"
|     <table>

#data
x<table>x
#errors
(1,1): expected-doctype-but-got-chars
(1,9): foster-parenting-character
(1,9): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     "xx"
|     <table>

#data
x<table><table>x
#errors
(1,1): expected-doctype-but-got-chars
(1,15): unexpected-start-tag-implies-end-tag
(1,16): foster-parenting-character
(1,16): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|     "x"
|     <table>

#data
<b>a<div></div><div></b>y
#errors
(1,3): expected-doctype-but-got-start-tag
(1,24): adoption-agency-1.3
(1,25): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "a"
|       <div>
|     <div>
|       <b>
|       "y"

#data
<a><div><p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,15): adoption-agency-1.3
(1,15): adoption-agency-1.3
(1,15): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|       <p>
|         <a>


#data
<!DOCTYPE html><math></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>

#data
<!DOCTYPE html><body><math></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>

#data
<!DOCTYPE html><math><mi>
#errors
(1,25) expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>

#data
<!DOCTYPE html><math><annotation-xml><svg><u>
#errors
(1,45) unexpected-html-element-in-foreign-content
(1,45) expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <svg svg>
|     <u>

#data
<!DOCTYPE html><body><select><math></math></select>
#errors
(1,35) unexpected-start-tag-in-select
(1,42) unexpected-end-tag-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE html><body><select><option><math></math></option></select>
#errors
(1,43) unexpected-start-tag-in-select
(1,50) unexpected-end-tag-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<!DOCTYPE html><body><table><math></math></table>
#errors
(1,34) unexpected-start-tag-implies-table-voodoo
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|     <table>

#data
<!DOCTYPE html><body><table><math><mi>foo</mi></math></table>
#errors
(1,34) foster-parenting-start-token
(1,39) foster-parenting-character
(1,40) foster-parenting-character
(1,41) foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|     <table>

#data
<!DOCTYPE html><body><table><math><mi>foo</mi><mi>bar</mi></math></table>
#errors
(1,34) foster-parenting-start-tag
(1,39) foster-parenting-character
(1,40) foster-parenting-character
(1,41) foster-parenting-character
(1,51) foster-parenting-character
(1,52) foster-parenting-character
(1,53) foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>

#data
<!DOCTYPE html><body><table><tbody><math><mi>foo</mi><mi>bar</mi></math></tbody></table>
#errors
(1,41) foster-parenting-start-tag
(1,46) foster-parenting-character
(1,47) foster-parenting-character
(1,48) foster-parenting-character
(1,58) foster-parenting-character
(1,59) foster-parenting-character
(1,60) foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>
|       <tbody>

#data
<!DOCTYPE html><body><table><tbody><tr><math><mi>foo</mi><mi>bar</mi></math></tr></tbody></table>
#errors
(1,45) foster-parenting-start-tag
(1,50) foster-parenting-character
(1,51) foster-parenting-character
(1,52) foster-parenting-character
(1,62) foster-parenting-character
(1,63) foster-parenting-character
(1,64) foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE html><body><table><tbody><tr><td><math><mi>foo</mi><mi>bar</mi></math></td></tr></tbody></table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math math>
|               <math mi>
|                 "foo"
|               <math mi>
|                 "bar"

#data
<!DOCTYPE html><body><table><tbody><tr><td><math><mi>foo</mi><mi>bar</mi></math><p>baz</td></tr></tbody></table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math math>
|               <math mi>
|                 "foo"
|               <math mi>
|                 "bar"
|             <p>
|               "baz"

#data
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi></math><p>baz</caption></table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|         <p>
|           "baz"

#data
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi><p>baz</table><p>quux
#errors
(1,70) unexpected-html-element-in-foreign-content
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|         <p>
|           "baz"
|     <p>
|       "quux"

#data
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi>baz</table><p>quux
#errors
(1,78) unexpected-end-tag
(1,78) expected-one-end-tag-but-got-another
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|           "baz"
|     <p>
|       "quux"

#data
<!DOCTYPE html><body><table><tr><td><select><math><mi>foo</mi><mi>bar</mi><p>baz</table><p>quux
#errors
(1,50) unexpected-start-tag-in-select
(1,54) unexpected-start-tag-in-select
(1,62) unexpected-end-tag-in-select
(1,66) unexpected-start-tag-in-select
(1,74) unexpected-end-tag-in-select
(1,77) unexpected-start-tag-in-select
(1,88) unexpected-table-element-end-tag-in-select-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "foobarbaz"
|     <p>
|       "quux"

#data
<!DOCTYPE html><body><table><select><math><mi>foo</mi><mi>bar</mi><p>baz</table><p>quux
#errors
(1,36) unexpected-start-tag-implies-table-voodoo
(1,42) unexpected-start-tag-in-select
(1,46) unexpected-start-tag-in-select
(1,54) unexpected-end-tag-in-select
(1,58) unexpected-start-tag-in-select
(1,66) unexpected-end-tag-in-select
(1,69) unexpected-start-tag-in-select
(1,80) unexpected-table-element-end-tag-in-select-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "foobarbaz"
|     <table>
|     <p>
|       "quux"

#data
<!DOCTYPE html><body></body></html><math><mi>foo</mi><mi>bar</mi><p>baz
#errors
(1,41) expected-eof-but-got-start-tag
(1,68) unexpected-html-element-in-foreign-content
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <p>
|       "baz"

#data
<!DOCTYPE html><body></body><math><mi>foo</mi><mi>bar</mi><p>baz
#errors
(1,34) unexpected-start-tag-after-body
(1,61) unexpected-html-element-in-foreign-content
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <p>
|       "baz"

#data
<!DOCTYPE html><frameset><math><mi></mi><mi></mi><p><span>
#errors
(1,31) unexpected-start-tag-in-frameset
(1,35) unexpected-start-tag-in-frameset
(1,40) unexpected-end-tag-in-frameset
(1,44) unexpected-start-tag-in-frameset
(1,49) unexpected-end-tag-in-frameset
(1,52) unexpected-start-tag-in-frameset
(1,58) unexpected-start-tag-in-frameset
(1,58) eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><body xlink:href=foo><math xlink:href=foo></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     <math math>
|       xlink href="foo"

#data
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo></mi></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"

#data
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo /></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"

#data
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo />bar</math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"
|       "bar"

#data
<!DOCTYPE html><svg></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<!DOCTYPE html><svg></svg><![CDATA[a]]>
#errors
(1,28) expected-dashes-or-doctype
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <!-- [CDATA[a]] -->

#data
<!DOCTYPE html><body><svg></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<!DOCTYPE html><body><select><svg></svg></select>
#errors
(1,34) unexpected-start-tag-in-select
(1,40) unexpected-end-tag-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE html><body><select><option><svg></svg></option></select>
#errors
(1,42) unexpected-start-tag-in-select
(1,48) unexpected-end-tag-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<!DOCTYPE html><body><table><svg></svg></table>
#errors
(1,33) foster-parenting-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <table>

#data
<!DOCTYPE html><body><table><svg><g>foo</g></svg></table>
#errors
(1,33) foster-parenting-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|     <table>

#data
<!DOCTYPE html><body><table><svg><g>foo</g><g>bar</g></svg></table>
#errors
(1,33) foster-parenting-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <table>

#data
<!DOCTYPE html><body><table><tbody><svg><g>foo</g><g>bar</g></svg></tbody></table>
#errors
(1,40) foster-parenting-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <table>
|       <tbody>

#data
<!DOCTYPE html><body><table><tbody><tr><svg><g>foo</g><g>bar</g></svg></tr></tbody></table>
#errors
(1,44) foster-parenting-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE html><body><table><tbody><tr><td><svg><g>foo</g><g>bar</g></svg></td></tr></tbody></table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "foo"
|               <svg g>
|                 "bar"

#data
<!DOCTYPE html><body><table><tbody><tr><td><svg><g>foo</g><g>bar</g></svg><p>baz</td></tr></tbody></table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "foo"
|               <svg g>
|                 "bar"
|             <p>
|               "baz"

#data
<!DOCTYPE html><body><table><caption><svg><g>foo</g><g>bar</g></svg><p>baz</caption></table>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|         <p>
|           "baz"

#data
<!DOCTYPE html><body><table><caption><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
#errors
(1,65) unexpected-html-element-in-foreign-content
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|         <p>
|           "baz"
|     <p>
|       "quux"

#data
<!DOCTYPE html><body><table><caption><svg><g>foo</g><g>bar</g>baz</table><p>quux
#errors
(1,73) unexpected-end-tag
(1,73) expected-one-end-tag-but-got-another
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|           "baz"
|     <p>
|       "quux"

#data
<!DOCTYPE html><body><table><colgroup><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
#errors
(1,43) foster-parenting-start-tag svg
(1,66) unexpected HTML-like start tag token in foreign content
(1,66) foster-parenting-start-tag
(1,67) foster-parenting-character
(1,68) foster-parenting-character
(1,69) foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"
|     <table>
|       <colgroup>
|     <p>
|       "quux"

#data
<!DOCTYPE html><body><table><tr><td><select><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
#errors
(1,49) unexpected-start-tag-in-select
(1,52) unexpected-start-tag-in-select
(1,59) unexpected-end-tag-in-select
(1,62) unexpected-start-tag-in-select
(1,69) unexpected-end-tag-in-select
(1,72) unexpected-start-tag-in-select
(1,83) unexpected-table-element-end-tag-in-select-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "foobarbaz"
|     <p>
|       "quux"

#data
<!DOCTYPE html><body><table><select><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
#errors
(1,36) unexpected-start-tag-implies-table-voodoo
(1,41) unexpected-start-tag-in-select
(1,44) unexpected-start-tag-in-select
(1,51) unexpected-end-tag-in-select
(1,54) unexpected-start-tag-in-select
(1,61) unexpected-end-tag-in-select
(1,64) unexpected-start-tag-in-select
(1,75) unexpected-table-element-end-tag-in-select-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "foobarbaz"
|     <table>
|     <p>
|       "quux"

#data
<!DOCTYPE html><body></body></html><svg><g>foo</g><g>bar</g><p>baz
#errors
(1,40) expected-eof-but-got-start-tag
(1,63) unexpected-html-element-in-foreign-content
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"

#data
<!DOCTYPE html><body></body><svg><g>foo</g><g>bar</g><p>baz
#errors
(1,33) unexpected-start-tag-after-body
(1,56) unexpected-html-element-in-foreign-content
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"

#data
<!DOCTYPE html><frameset><svg><g></g><g></g><p><span>
#errors
(1,30) unexpected-start-tag-in-frameset
(1,33) unexpected-start-tag-in-frameset
(1,37) unexpected-end-tag-in-frameset
(1,40) unexpected-start-tag-in-frameset
(1,44) unexpected-end-tag-in-frameset
(1,47) unexpected-start-tag-in-frameset
(1,53) unexpected-start-tag-in-frameset
(1,53) eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><frameset></frameset><svg><g></g><g></g><p><span>
#errors
(1,41) unexpected-start-tag-after-frameset
(1,44) unexpected-start-tag-after-frameset
(1,48) unexpected-end-tag-after-frameset
(1,51) unexpected-start-tag-after-frameset
(1,55) unexpected-end-tag-after-frameset
(1,58) unexpected-start-tag-after-frameset
(1,64) unexpected-start-tag-after-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><body xlink:href=foo><svg xlink:href=foo></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     <svg svg>
|       xlink href="foo"

#data
<!DOCTYPE html><body xlink:href=foo xml:lang=en><svg><g xml:lang=en xlink:href=foo></g></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg svg>
|       <svg g>
|         xlink href="foo"
|         xml lang="en"

#data
<!DOCTYPE html><body xlink:href=foo xml:lang=en><svg><g xml:lang=en xlink:href=foo /></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg svg>
|       <svg g>
|         xlink href="foo"
|         xml lang="en"

#data
<!DOCTYPE html><body xlink:href=foo xml:lang=en><svg><g xml:lang=en xlink:href=foo />bar</svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg svg>
|       <svg g>
|         xlink href="foo"
|         xml lang="en"
|       "bar"

#data
<svg></path>
#errors
(1,5) expected-doctype-but-got-start-tag
(1,12) unexpected-end-tag
(1,12) unexpected-end-tag
(1,12) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<div><svg></div>a
#errors
(1,5) expected-doctype-but-got-start-tag
(1,16) unexpected-end-tag
(1,16) end-tag-too-early
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <svg svg>
|     "a"

#data
<div><svg><path></div>a
#errors
(1,5) expected-doctype-but-got-start-tag
(1,22) unexpected-end-tag
(1,22) end-tag-too-early
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <svg svg>
|         <svg path>
|     "a"

#data
<div><svg><path></svg><path>
#errors
(1,5) expected-doctype-but-got-start-tag
(1,22) unexpected-end-tag
(1,28) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <svg svg>
|         <svg path>
|       <path>

#data
<div><svg><path><foreignObject><p></div>a
#errors
(1,5) expected-doctype-but-got-start-tag
(1,40) end-tag-too-early
(1,41) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <svg svg>
|         <svg path>
|           <svg foreignObject>
|             <p>
|               "a"

#data
<!DOCTYPE html><svg><desc><div><svg><ul>a
#errors
(1,40) unexpected-html-element-in-foreign-content
(1,41) expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg desc>
|         <div>
|           <svg svg>
|           <ul>
|             "a"

#data
<!DOCTYPE html><svg><desc><svg><ul>a
#errors
(1,35) unexpected-html-element-in-foreign-content
(1,36) expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg desc>
|         <svg svg>
|         <ul>
|           "a"

#data
<!DOCTYPE html><p><svg><desc><p>
#errors
(1,32) expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg desc>
|           <p>

#data
<!DOCTYPE html><p><svg><title><p>
#errors
(1,33) expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <svg svg>
|         <svg title>
|           <p>

#data
<div><svg><path><foreignObject><p></foreignObject><p>
#errors
(1,5) expected-doctype-but-got-start-tag
(1,50) unexpected-end-tag
(1,53) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <svg svg>
|         <svg path>
|           <svg foreignObject>
|             <p>
|             <p>

#data
<math><mi><div><object><div><span></span></div></object></div></mi><mi>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,71) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <div>
|           <object>
|             <div>
|               <span>
|       <math mi>

#data
<math><mi><svg><foreignObject><div><div></div></div></foreignObject></svg></mi><mi>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,83) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <svg svg>
|           <svg foreignObject>
|             <div>
|               <div>
|       <math mi>

#data
<svg><script></script><path>
#errors
(1,5) expected-doctype-but-got-start-tag
(1,28) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg script>
|       <svg path>

#data
<table><svg></svg><tr>
#errors
(1,7) expected-doctype-but-got-start-tag
(1,12) unexpected-start-tag-implies-table-voodoo
(1,22) eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <table>
|       <tbody>
|         <tr>

#data
<math><mi><mglyph>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,18) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <math mglyph>

#data
<math><mi><malignmark>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,22) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         <math malignmark>

#data
<math><mo><mglyph>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,18) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mo>
|         <math mglyph>

#data
<math><mo><malignmark>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,22) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mo>
|         <math malignmark>

#data
<math><mn><mglyph>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,18) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mn>
|         <math mglyph>

#data
<math><mn><malignmark>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,22) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mn>
|         <math malignmark>

#data
<math><ms><mglyph>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,18) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math ms>
|         <math mglyph>

#data
<math><ms><malignmark>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,22) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math ms>
|         <math malignmark>

#data
<math><mtext><mglyph>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,21) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mtext>
|         <math mglyph>

#data
<math><mtext><malignmark>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,25) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mtext>
|         <math malignmark>

#data
<math><annotation-xml><svg></svg></annotation-xml><mi>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,54) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <svg svg>
|       <math mi>

#data
<math><annotation-xml><svg><foreignObject><div><math><mi></mi></math><span></span></div></foreignObject><path></path></svg></annotation-xml><mi>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,144) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <svg svg>
|           <svg foreignObject>
|             <div>
|               <math math>
|                 <math mi>
|               <span>
|           <svg path>
|       <math mi>

#data
<math><annotation-xml><svg><foreignObject><math><mi><svg></svg></mi><mo></mo></math><span></span></foreignObject><path></path></svg></annotation-xml><mi>
#errors
(1,6) expected-doctype-but-got-start-tag
(1,153) expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <svg svg>
|           <svg foreignObject>
|             <math math>
|               <math mi>
|                 <svg svg>
|               <math mo>
|             <span>
|           <svg path>
|       <math mi>

#data
<!DOCTYPE html><body><svg attributeName='' attributeType='' baseFrequency='' baseProfile='' calcMode='' clipPathUnits='' diffuseConstant='' edgeMode='' filterUnits='' glyphRef='' gradientTransform='' gradientUnits='' kernelMatrix='' kernelUnitLength='' keyPoints='' keySplines='' keyTimes='' lengthAdjust='' limitingConeAngle='' markerHeight='' markerUnits='' markerWidth='' maskContentUnits='' maskUnits='' numOctaves='' pathLength='' patternContentUnits='' patternTransform='' patternUnits='' pointsAtX='' pointsAtY='' pointsAtZ='' preserveAlpha='' preserveAspectRatio='' primitiveUnits='' refX='' refY='' repeatCount='' repeatDur='' requiredExtensions='' requiredFeatures='' specularConstant='' specularExponent='' spreadMethod='' startOffset='' stdDeviation='' stitchTiles='' surfaceScale='' systemLanguage='' tableValues='' targetX='' targetY='' textLength='' viewBox='' viewTarget='' xChannelSelector='' yChannelSelector='' zoomAndPan=''></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       attributeName=""
|       attributeType=""
|       baseFrequency=""
|       baseProfile=""
|       calcMode=""
|       clipPathUnits=""
|       diffuseConstant=""
|       edgeMode=""
|       filterUnits=""
|       glyphRef=""
|       gradientTransform=""
|       gradientUnits=""
|       kernelMatrix=""
|       kernelUnitLength=""
|       keyPoints=""
|       keySplines=""
|       keyTimes=""
|       lengthAdjust=""
|       limitingConeAngle=""
|       markerHeight=""
|       markerUnits=""
|       markerWidth=""
|       maskContentUnits=""
|       maskUnits=""
|       numOctaves=""
|       pathLength=""
|       patternContentUnits=""
|       patternTransform=""
|       patternUnits=""
|       pointsAtX=""
|       pointsAtY=""
|       pointsAtZ=""
|       preserveAlpha=""
|       preserveAspectRatio=""
|       primitiveUnits=""
|       refX=""
|       refY=""
|       repeatCount=""
|       repeatDur=""
|       requiredExtensions=""
|       requiredFeatures=""
|       specularConstant=""
|       specularExponent=""
|       spreadMethod=""
|       startOffset=""
|       stdDeviation=""
|       stitchTiles=""
|       surfaceScale=""
|       systemLanguage=""
|       tableValues=""
|       targetX=""
|       targetY=""
|       textLength=""
|       viewBox=""
|       viewTarget=""
|       xChannelSelector=""
|       yChannelSelector=""
|       zoomAndPan=""

#data
<!DOCTYPE html><BODY><SVG ATTRIBUTENAME='' ATTRIBUTETYPE='' BASEFREQUENCY='' BASEPROFILE='' CALCMODE='' CLIPPATHUNITS='' DIFFUSECONSTANT='' EDGEMODE='' FILTERUNITS='' GLYPHREF='' GRADIENTTRANSFORM='' GRADIENTUNITS='' KERNELMATRIX='' KERNELUNITLENGTH='' KEYPOINTS='' KEYSPLINES='' KEYTIMES='' LENGTHADJUST='' LIMITINGCONEANGLE='' MARKERHEIGHT='' MARKERUNITS='' MARKERWIDTH='' MASKCONTENTUNITS='' MASKUNITS='' NUMOCTAVES='' PATHLENGTH='' PATTERNCONTENTUNITS='' PATTERNTRANSFORM='' PATTERNUNITS='' POINTSATX='' POINTSATY='' POINTSATZ='' PRESERVEALPHA='' PRESERVEASPECTRATIO='' PRIMITIVEUNITS='' REFX='' REFY='' REPEATCOUNT='' REPEATDUR='' REQUIREDEXTENSIONS='' REQUIREDFEATURES='' SPECULARCONSTANT='' SPECULAREXPONENT='' SPREADMETHOD='' STARTOFFSET='' STDDEVIATION='' STITCHTILES='' SURFACESCALE='' SYSTEMLANGUAGE='' TABLEVALUES='' TARGETX='' TARGETY='' TEXTLENGTH='' VIEWBOX='' VIEWTARGET='' XCHANNELSELECTOR='' YCHANNELSELECTOR='' ZOOMANDPAN=''></SVG>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       attributeName=""
|       attributeType=""
|       baseFrequency=""
|       baseProfile=""
|       calcMode=""
|       clipPathUnits=""
|       diffuseConstant=""
|       edgeMode=""
|       filterUnits=""
|       glyphRef=""
|       gradientTransform=""
|       gradientUnits=""
|       kernelMatrix=""
|       kernelUnitLength=""
|       keyPoints=""
|       keySplines=""
|       keyTimes=""
|       lengthAdjust=""
|       limitingConeAngle=""
|       markerHeight=""
|       markerUnits=""
|       markerWidth=""
|       maskContentUnits=""
|       maskUnits=""
|       numOctaves=""
|       pathLength=""
|       patternContentUnits=""
|       patternTransform=""
|       patternUnits=""
|       pointsAtX=""
|       pointsAtY=""
|       pointsAtZ=""
|       preserveAlpha=""
|       preserveAspectRatio=""
|       primitiveUnits=""
|       refX=""
|       refY=""
|       repeatCount=""
|       repeatDur=""
|       requiredExtensions=""
|       requiredFeatures=""
|       specularConstant=""
|       specularExponent=""
|       spreadMethod=""
|       startOffset=""
|       stdDeviation=""
|       stitchTiles=""
|       surfaceScale=""
|       systemLanguage=""
|       tableValues=""
|       targetX=""
|       targetY=""
|       textLength=""
|       viewBox=""
|       viewTarget=""
|       xChannelSelector=""
|       yChannelSelector=""
|       zoomAndPan=""

#data
<!DOCTYPE html><body><svg attributename='' attributetype='' basefrequency='' baseprofile='' calcmode='' clippathunits='' diffuseconstant='' edgemode='' filterunits='' filterres='' glyphref='' gradienttransform='' gradientunits='' kernelmatrix='' kernelunitlength='' keypoints='' keysplines='' keytimes='' lengthadjust='' limitingconeangle='' markerheight='' markerunits='' markerwidth='' maskcontentunits='' maskunits='' numoctaves='' pathlength='' patterncontentunits='' patterntransform='' patternunits='' pointsatx='' pointsaty='' pointsatz='' preservealpha='' preserveaspectratio='' primitiveunits='' refx='' refy='' repeatcount='' repeatdur='' requiredextensions='' requiredfeatures='' specularconstant='' specularexponent='' spreadmethod='' startoffset='' stddeviation='' stitchtiles='' surfacescale='' systemlanguage='' tablevalues='' targetx='' targety='' textlength='' viewbox='' viewtarget='' xchannelselector='' ychannelselector='' zoomandpan=''></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       attributeName=""
|       attributeType=""
|       baseFrequency=""
|       baseProfile=""
|       calcMode=""
|       clipPathUnits=""
|       diffuseConstant=""
|       edgeMode=""
|       filterUnits=""
|       filterres=""
|       glyphRef=""
|       gradientTransform=""
|       gradientUnits=""
|       kernelMatrix=""
|       kernelUnitLength=""
|       keyPoints=""
|       keySplines=""
|       keyTimes=""
|       lengthAdjust=""
|       limitingConeAngle=""
|       markerHeight=""
|       markerUnits=""
|       markerWidth=""
|       maskContentUnits=""
|       maskUnits=""
|       numOctaves=""
|       pathLength=""
|       patternContentUnits=""
|       patternTransform=""
|       patternUnits=""
|       pointsAtX=""
|       pointsAtY=""
|       pointsAtZ=""
|       preserveAlpha=""
|       preserveAspectRatio=""
|       primitiveUnits=""
|       refX=""
|       refY=""
|       repeatCount=""
|       repeatDur=""
|       requiredExtensions=""
|       requiredFeatures=""
|       specularConstant=""
|       specularExponent=""
|       spreadMethod=""
|       startOffset=""
|       stdDeviation=""
|       stitchTiles=""
|       surfaceScale=""
|       systemLanguage=""
|       tableValues=""
|       targetX=""
|       targetY=""
|       textLength=""
|       viewBox=""
|       viewTarget=""
|       xChannelSelector=""
|       yChannelSelector=""
|       zoomAndPan=""

#data
<!DOCTYPE html><body><math attributeName='' attributeType='' baseFrequency='' baseProfile='' calcMode='' clipPathUnits='' diffuseConstant='' edgeMode='' filterUnits='' glyphRef='' gradientTransform='' gradientUnits='' kernelMatrix='' kernelUnitLength='' keyPoints='' keySplines='' keyTimes='' lengthAdjust='' limitingConeAngle='' markerHeight='' markerUnits='' markerWidth='' maskContentUnits='' maskUnits='' numOctaves='' pathLength='' patternContentUnits='' patternTransform='' patternUnits='' pointsAtX='' pointsAtY='' pointsAtZ='' preserveAlpha='' preserveAspectRatio='' primitiveUnits='' refX='' refY='' repeatCount='' repeatDur='' requiredExtensions='' requiredFeatures='' specularConstant='' specularExponent='' spreadMethod='' startOffset='' stdDeviation='' stitchTiles='' surfaceScale='' systemLanguage='' tableValues='' targetX='' targetY='' textLength='' viewBox='' viewTarget='' xChannelSelector='' yChannelSelector='' zoomAndPan=''></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       attributename=""
|       attributetype=""
|       basefrequency=""
|       baseprofile=""
|       calcmode=""
|       clippathunits=""
|       diffuseconstant=""
|       edgemode=""
|       filterunits=""
|       glyphref=""
|       gradienttransform=""
|       gradientunits=""
|       kernelmatrix=""
|       kernelunitlength=""
|       keypoints=""
|       keysplines=""
|       keytimes=""
|       lengthadjust=""
|       limitingconeangle=""
|       markerheight=""
|       markerunits=""
|       markerwidth=""
|       maskcontentunits=""
|       maskunits=""
|       numoctaves=""
|       pathlength=""
|       patterncontentunits=""
|       patterntransform=""
|       patternunits=""
|       pointsatx=""
|       pointsaty=""
|       pointsatz=""
|       preservealpha=""
|       preserveaspectratio=""
|       primitiveunits=""
|       refx=""
|       refy=""
|       repeatcount=""
|       repeatdur=""
|       requiredextensions=""
|       requiredfeatures=""
|       specularconstant=""
|       specularexponent=""
|       spreadmethod=""
|       startoffset=""
|       stddeviation=""
|       stitchtiles=""
|       surfacescale=""
|       systemlanguage=""
|       tablevalues=""
|       targetx=""
|       targety=""
|       textlength=""
|       viewbox=""
|       viewtarget=""
|       xchannelselector=""
|       ychannelselector=""
|       zoomandpan=""

#data
<!DOCTYPE html><body><svg><altGlyph /><altGlyphDef /><altGlyphItem /><animateColor /><animateMotion /><animateTransform /><clipPath /><feBlend /><feColorMatrix /><feComponentTransfer /><feComposite /><feConvolveMatrix /><feDiffuseLighting /><feDisplacementMap /><feDistantLight /><feFlood /><feFuncA /><feFuncB /><feFuncG /><feFuncR /><feGaussianBlur /><feImage /><feMerge /><feMergeNode /><feMorphology /><feOffset /><fePointLight /><feSpecularLighting /><feSpotLight /><feTile /><feTurbulence /><foreignObject /><glyphRef /><linearGradient /><radialGradient /><textPath /></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg altGlyph>
|       <svg altGlyphDef>
|       <svg altGlyphItem>
|       <svg animateColor>
|       <svg animateMotion>
|       <svg animateTransform>
|       <svg clipPath>
|       <svg feBlend>
|       <svg feColorMatrix>
|       <svg feComponentTransfer>
|       <svg feComposite>
|       <svg feConvolveMatrix>
|       <svg feDiffuseLighting>
|       <svg feDisplacementMap>
|       <svg feDistantLight>
|       <svg feFlood>
|       <svg feFuncA>
|       <svg feFuncB>
|       <svg feFuncG>
|       <svg feFuncR>
|       <svg feGaussianBlur>
|       <svg feImage>
|       <svg feMerge>
|       <svg feMergeNode>
|       <svg feMorphology>
|       <svg feOffset>
|       <svg fePointLight>
|       <svg feSpecularLighting>
|       <svg feSpotLight>
|       <svg feTile>
|       <svg feTurbulence>
|       <svg foreignObject>
|       <svg glyphRef>
|       <svg linearGradient>
|       <svg radialGradient>
|       <svg textPath>

#data
<!DOCTYPE html><body><svg><altglyph /><altglyphdef /><altglyphitem /><animatecolor /><animatemotion /><animatetransform /><clippath /><feblend /><fecolormatrix /><fecomponenttransfer /><fecomposite /><feconvolvematrix /><fediffuselighting /><fedisplacementmap /><fedistantlight /><feflood /><fefunca /><fefuncb /><fefuncg /><fefuncr /><fegaussianblur /><feimage /><femerge /><femergenode /><femorphology /><feoffset /><fepointlight /><fespecularlighting /><fespotlight /><fetile /><feturbulence /><foreignobject /><glyphref /><lineargradient /><radialgradient /><textpath /></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg altGlyph>
|       <svg altGlyphDef>
|       <svg altGlyphItem>
|       <svg animateColor>
|       <svg animateMotion>
|       <svg animateTransform>
|       <svg clipPath>
|       <svg feBlend>
|       <svg feColorMatrix>
|       <svg feComponentTransfer>
|       <svg feComposite>
|       <svg feConvolveMatrix>
|       <svg feDiffuseLighting>
|       <svg feDisplacementMap>
|       <svg feDistantLight>
|       <svg feFlood>
|       <svg feFuncA>
|       <svg feFuncB>
|       <svg feFuncG>
|       <svg feFuncR>
|       <svg feGaussianBlur>
|       <svg feImage>
|       <svg feMerge>
|       <svg feMergeNode>
|       <svg feMorphology>
|       <svg feOffset>
|       <svg fePointLight>
|       <svg feSpecularLighting>
|       <svg feSpotLight>
|       <svg feTile>
|       <svg feTurbulence>
|       <svg foreignObject>
|       <svg glyphRef>
|       <svg linearGradient>
|       <svg radialGradient>
|       <svg textPath>

#data
<!DOCTYPE html><BODY><SVG><ALTGLYPH /><ALTGLYPHDEF /><ALTGLYPHITEM /><ANIMATECOLOR /><ANIMATEMOTION /><ANIMATETRANSFORM /><CLIPPATH /><FEBLEND /><FECOLORMATRIX /><FECOMPONENTTRANSFER /><FECOMPOSITE /><FECONVOLVEMATRIX /><FEDIFFUSELIGHTING /><FEDISPLACEMENTMAP /><FEDISTANTLIGHT /><FEFLOOD /><FEFUNCA /><FEFUNCB /><FEFUNCG /><FEFUNCR /><FEGAUSSIANBLUR /><FEIMAGE /><FEMERGE /><FEMERGENODE /><FEMORPHOLOGY /><FEOFFSET /><FEPOINTLIGHT /><FESPECULARLIGHTING /><FESPOTLIGHT /><FETILE /><FETURBULENCE /><FOREIGNOBJECT /><GLYPHREF /><LINEARGRADIENT /><RADIALGRADIENT /><TEXTPATH /></SVG>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg altGlyph>
|       <svg altGlyphDef>
|       <svg altGlyphItem>
|       <svg animateColor>
|       <svg animateMotion>
|       <svg animateTransform>
|       <svg clipPath>
|       <svg feBlend>
|       <svg feColorMatrix>
|       <svg feComponentTransfer>
|       <svg feComposite>
|       <svg feConvolveMatrix>
|       <svg feDiffuseLighting>
|       <svg feDisplacementMap>
|       <svg feDistantLight>
|       <svg feFlood>
|       <svg feFuncA>
|       <svg feFuncB>
|       <svg feFuncG>
|       <svg feFuncR>
|       <svg feGaussianBlur>
|       <svg feImage>
|       <svg feMerge>
|       <svg feMergeNode>
|       <svg feMorphology>
|       <svg feOffset>
|       <svg fePointLight>
|       <svg feSpecularLighting>
|       <svg feSpotLight>
|       <svg feTile>
|       <svg feTurbulence>
|       <svg foreignObject>
|       <svg glyphRef>
|       <svg linearGradient>
|       <svg radialGradient>
|       <svg textPath>

#data
<!DOCTYPE html><body><math><altGlyph /><altGlyphDef /><altGlyphItem /><animateColor /><animateMotion /><animateTransform /><clipPath /><feBlend /><feColorMatrix /><feComponentTransfer /><feComposite /><feConvolveMatrix /><feDiffuseLighting /><feDisplacementMap /><feDistantLight /><feFlood /><feFuncA /><feFuncB /><feFuncG /><feFuncR /><feGaussianBlur /><feImage /><feMerge /><feMergeNode /><feMorphology /><feOffset /><fePointLight /><feSpecularLighting /><feSpotLight /><feTile /><feTurbulence /><foreignObject /><glyphRef /><linearGradient /><radialGradient /><textPath /></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math altglyph>
|       <math altglyphdef>
|       <math altglyphitem>
|       <math animatecolor>
|       <math animatemotion>
|       <math animatetransform>
|       <math clippath>
|       <math feblend>
|       <math fecolormatrix>
|       <math fecomponenttransfer>
|       <math fecomposite>
|       <math feconvolvematrix>
|       <math fediffuselighting>
|       <math fedisplacementmap>
|       <math fedistantlight>
|       <math feflood>
|       <math fefunca>
|       <math fefuncb>
|       <math fefuncg>
|       <math fefuncr>
|       <math fegaussianblur>
|       <math feimage>
|       <math femerge>
|       <math femergenode>
|       <math femorphology>
|       <math feoffset>
|       <math fepointlight>
|       <math fespecularlighting>
|       <math fespotlight>
|       <math fetile>
|       <math feturbulence>
|       <math foreignobject>
|       <math glyphref>
|       <math lineargradient>
|       <math radialgradient>
|       <math textpath>

#data
<!DOCTYPE html><body><svg><solidColor /></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg solidcolor>

#data
<!DOCTYPE html><p><b><i><u></p> <p>X
#errors
(1,31): unexpected-end-tag
(1,36): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         <i>
|           <u>
|     <b>
|       <i>
|         <u>
|           " "
|           <p>
|             "X"

#data
<p><b><i><u></p>
<p>X
#errors
(1,3): expected-doctype-but-got-start-tag
(1,16): unexpected-end-tag
(2,4): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         <i>
|           <u>
|     <b>
|       <i>
|         <u>
|           "
"
|           <p>
|             "X"

#data
<!doctype html></html> <head>
#errors
(1,29): expected-eof-but-got-start-tag
(1,29): unexpected-start-tag-ignored
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " "

#data
<!doctype html></body><meta>
#errors
(1,28): unexpected-start-tag-after-body
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <meta>

#data
<html></html><!-- foo -->
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
| <!--  foo  -->

#data
<!doctype html></body><title>X</title>
#errors
(1,29): unexpected-start-tag-after-body
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <title>
|       "X"

#data
<!doctype html><table> X<meta></table>
#errors
(1,23): foster-parenting-character
(1,24): foster-parenting-character
(1,30): foster-parenting-start-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " X"
|     <meta>
|     <table>

#data
<!doctype html><table> x</table>
#errors
(1,23): foster-parenting-character
(1,24): foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " x"
|     <table>

#data
<!doctype html><table> x </table>
#errors
(1,23): foster-parenting-character
(1,24): foster-parenting-character
(1,25): foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " x "
|     <table>

#data
<!doctype html><table><tr> x</table>
#errors
(1,27): foster-parenting-character
(1,28): foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " x"
|     <table>
|       <tbody>
|         <tr>

#data
<!doctype html><table>X<style> <tr>x </style> </table>
#errors
(1,23): foster-parenting-character
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"
|     <table>
|       <style>
|         " <tr>x "
|       " "

#data
<!doctype html><div><table><a>foo</a> <tr><td>bar</td> </tr></table></div>
#errors
(1,30): foster-parenting-start-tag
(1,31): foster-parenting-character
(1,32): foster-parenting-character
(1,33): foster-parenting-character
(1,37): foster-parenting-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>
|       <a>
|         "foo"
|       <table>
|         " "
|         <tbody>
|           <tr>
|             <td>
|               "bar"
|             " "

#data
<frame></frame></frame><frameset><frame><frameset><frame></frameset><noframes></frameset><noframes>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,7): unexpected-start-tag-ignored
(1,15): unexpected-end-tag
(1,23): unexpected-end-tag
(1,33): unexpected-start-tag
(1,99): expected-named-closing-tag-but-got-eof
(1,99): eof-in-frameset
#document
| <html>
|   <head>
|   <frameset>
|     <frame>
|     <frameset>
|       <frame>
|     <noframes>
|       "</frameset><noframes>"

#data
<!DOCTYPE html><object></html>
#errors
(1,30): expected-body-in-scope
(1,30): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <object>

#data
<!doctype html><script>
#errors
(1,23): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|   <body>

#data
<!doctype html><script>a
#errors
(1,24): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "a"
|   <body>

#data
<!doctype html><script><
#errors
(1,24): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<"
|   <body>

#data
<!doctype html><script></
#errors
(1,25): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</"
|   <body>

#data
<!doctype html><script></S
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</S"
|   <body>

#data
<!doctype html><script></SC
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SC"
|   <body>

#data
<!doctype html><script></SCR
#errors
(1,28): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SCR"
|   <body>

#data
<!doctype html><script></SCRI
#errors
(1,29): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SCRI"
|   <body>

#data
<!doctype html><script></SCRIP
#errors
(1,30): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SCRIP"
|   <body>

#data
<!doctype html><script></SCRIPT
#errors
(1,31): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</SCRIPT"
|   <body>

#data
<!doctype html><script></SCRIPT 
#errors
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|   <body>

#data
<!doctype html><script></s
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</s"
|   <body>

#data
<!doctype html><script></sc
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</sc"
|   <body>

#data
<!doctype html><script></scr
#errors
(1,28): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</scr"
|   <body>

#data
<!doctype html><script></scri
#errors
(1,29): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</scri"
|   <body>

#data
<!doctype html><script></scrip
#errors
(1,30): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</scrip"
|   <body>

#data
<!doctype html><script></script
#errors
(1,31): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "</script"
|   <body>

#data
<!doctype html><script></script 
#errors
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|   <body>

#data
<!doctype html><script><!
#errors
(1,25): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!"
|   <body>

#data
<!doctype html><script><!a
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!a"
|   <body>

#data
<!doctype html><script><!-
#errors
(1,26): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!-"
|   <body>

#data
<!doctype html><script><!-a
#errors
(1,27): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!-a"
|   <body>

#data
<!doctype html><script><!--
#errors
(1,27): expected-named-closing-tag-but-got-eof
(1,27): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--"
|   <body>

#data
<!doctype html><script><!--a
#errors
(1,28): expected-named-closing-tag-but-got-eof
(1,28): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--a"
|   <body>

#data
<!doctype html><script><!--<
#errors
(1,28): expected-named-closing-tag-but-got-eof
(1,28): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<"
|   <body>

#data
<!doctype html><script><!--<a
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<a"
|   <body>

#data
<!doctype html><script><!--</
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--</"
|   <body>

#data
<!doctype html><script><!--</script
#errors
(1,35): expected-named-closing-tag-but-got-eof
(1,35): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--</script"
|   <body>

#data
<!doctype html><script><!--</script 
#errors
(1,36): expected-attribute-name-but-got-eof
(1,36): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--"
|   <body>

#data
<!doctype html><script><!--<s
#errors
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<s"
|   <body>

#data
<!doctype html><script><!--<script
#errors
(1,34): expected-named-closing-tag-but-got-eof
(1,34): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script"
|   <body>

#data
<!doctype html><script><!--<script 
#errors
(1,35): eof-in-script-in-script
(1,35): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script "
|   <body>

#data
<!doctype html><script><!--<script <
#errors
(1,36): eof-in-script-in-script
(1,36): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script <"
|   <body>

#data
<!doctype html><script><!--<script <a
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script <a"
|   <body>

#data
<!doctype html><script><!--<script </
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </"
|   <body>

#data
<!doctype html><script><!--<script </s
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </s"
|   <body>

#data
<!doctype html><script><!--<script </script
#errors
(1,43): eof-in-script-in-script
(1,43): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script"
|   <body>

#data
<!doctype html><script><!--<script </scripta
#errors
(1,44): eof-in-script-in-script
(1,44): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </scripta"
|   <body>

#data
<!doctype html><script><!--<script </script 
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<!doctype html><script><!--<script </script>
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script>"
|   <body>

#data
<!doctype html><script><!--<script </script/
#errors
(1,44): expected-named-closing-tag-but-got-eof
(1,44): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script/"
|   <body>

#data
<!doctype html><script><!--<script </script <
#errors
(1,45): expected-named-closing-tag-but-got-eof
(1,45): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script <"
|   <body>

#data
<!doctype html><script><!--<script </script <a
#errors
(1,46): expected-named-closing-tag-but-got-eof
(1,46): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script <a"
|   <body>

#data
<!doctype html><script><!--<script </script </
#errors
(1,46): expected-named-closing-tag-but-got-eof
(1,46): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script </"
|   <body>

#data
<!doctype html><script><!--<script </script </script
#errors
(1,52): expected-named-closing-tag-but-got-eof
(1,52): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script </script"
|   <body>

#data
<!doctype html><script><!--<script </script </script 
#errors
(1,53): expected-attribute-name-but-got-eof
(1,53): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<!doctype html><script><!--<script </script </script/
#errors
(1,53): unexpected-EOF-after-solidus-in-tag
(1,53): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<!doctype html><script><!--<script </script </script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<!doctype html><script><!--<script -
#errors
(1,36): eof-in-script-in-script
(1,36): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -"
|   <body>

#data
<!doctype html><script><!--<script -a
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -a"
|   <body>

#data
<!doctype html><script><!--<script -<
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -<"
|   <body>

#data
<!doctype html><script><!--<script --
#errors
(1,37): eof-in-script-in-script
(1,37): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --"
|   <body>

#data
<!doctype html><script><!--<script --a
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --a"
|   <body>

#data
<!doctype html><script><!--<script --<
#errors
(1,38): eof-in-script-in-script
(1,38): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --<"
|   <body>

#data
<!doctype html><script><!--<script -->
#errors
(1,38): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<!doctype html><script><!--<script --><
#errors
(1,39): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --><"
|   <body>

#data
<!doctype html><script><!--<script --></
#errors
(1,40): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --></"
|   <body>

#data
<!doctype html><script><!--<script --></script
#errors
(1,46): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script --></script"
|   <body>

#data
<!doctype html><script><!--<script --></script 
#errors
(1,47): expected-attribute-name-but-got-eof
(1,47): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<!doctype html><script><!--<script --></script/
#errors
(1,47): unexpected-EOF-after-solidus-in-tag
(1,47): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<!doctype html><script><!--<script --></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<!doctype html><script><!--<script><\/script>--></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script><\/script>-->"
|   <body>

#data
<!doctype html><script><!--<script></scr'+'ipt>--></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></scr'+'ipt>-->"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>--><!--</script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>--><!--"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>-- ></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>-- >"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>- -></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>- ->"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>- - ></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>- - >"
|   <body>

#data
<!doctype html><script><!--<script></script><script></script>-></script>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>->"
|   <body>

#data
<!doctype html><script><!--<script>--!></script>X
#errors
(1,49): expected-named-closing-tag-but-got-eof
(1,49): unexpected-EOF-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script>--!></script>X"
|   <body>

#data
<!doctype html><script><!--<scr'+'ipt></script>--></script>
#errors
(1,59): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<scr'+'ipt>"
|   <body>
|     "-->"

#data
<!doctype html><script><!--<script></scr'+'ipt></script>X
#errors
(1,57): expected-named-closing-tag-but-got-eof
(1,57): unexpected-eof-in-text-mode
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       "<!--<script></scr'+'ipt></script>X"
|   <body>

#data
<!doctype html><style><!--<style></style>--></style>
#errors
(1,52): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--<style>"
|   <body>
|     "-->"

#data
<!doctype html><style><!--</style>X
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--"
|   <body>
|     "X"

#data
<!doctype html><style><!--...</style>...--></style>
#errors
(1,51): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--..."
|   <body>
|     "...-->"

#data
<!doctype html><style><!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style></style>X
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style>"
|   <body>
|     "X"

#data
<!doctype html><style><!--...<style><!--...--!></style>--></style>
#errors
(1,66): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--...<style><!--...--!>"
|   <body>
|     "-->"

#data
<!doctype html><style><!--...</style><!-- --><style>@import ...</style>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "<!--..."
|     <!--   -->
|     <style>
|       "@import ..."
|   <body>

#data
<!doctype html><style>...<style><!--...</style><!-- --></style>
#errors
(1,63): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "...<style><!--..."
|     <!--   -->
|   <body>

#data
<!doctype html><style>...<!--[if IE]><style>...</style>X
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <style>
|       "...<!--[if IE]><style>..."
|   <body>
|     "X"

#data
<!doctype html><title><!--<title></title>--></title>
#errors
(1,52): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "<!--<title>"
|   <body>
|     "-->"

#data
<!doctype html><title>&lt;/title></title>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "</title>"
|   <body>

#data
<!doctype html><title>foo/title><link></head><body>X
#errors
(1,52): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "foo/title><link></head><body>X"
|   <body>

#data
<!doctype html><noscript><!--<noscript></noscript>--></noscript>
#errors
(1,64): unexpected-end-tag
#script-on
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       "<!--<noscript>"
|   <body>
|     "-->"

#data
<!doctype html><noscript><!--<noscript></noscript>--></noscript>
#errors
#script-off
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       <!-- <noscript></noscript> -->
|   <body>

#data
<!doctype html><noscript><!--</noscript>X<noscript>--></noscript>
#errors
#script-on
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       "<!--"
|   <body>
|     "X"
|     <noscript>
|       "-->"

#data
<!doctype html><noscript><!--</noscript>X<noscript>--></noscript>
#errors
#script-off
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       <!-- </noscript>X<noscript> -->
|   <body>

#data
<!doctype html><noscript><iframe></noscript>X
#errors
#script-on
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|       "<iframe>"
|   <body>
|     "X"

#data
<!doctype html><noscript><iframe></noscript>X
#errors
 * (1,34) unexpected token in head noscript
 * (1,46) unexpected EOF
#script-off
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noscript>
|   <body>
|     <iframe>
|       "</noscript>X"

#data
<!doctype html><noframes><!--<noframes></noframes>--></noframes>
#errors
(1,64): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noframes>
|       "<!--<noframes>"
|   <body>
|     "-->"

#data
<!doctype html><noframes><body><script><!--...</script></body></noframes></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <noframes>
|       "<body><script><!--...</script></body>"
|   <body>

#data
<!doctype html><textarea><!--<textarea></textarea>--></textarea>
#errors
(1,64): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<!--<textarea>"
|     "-->"

#data
<!doctype html><textarea>&lt;/textarea></textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "</textarea>"

#data
<!doctype html><textarea>&lt;</textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<"

#data
<!doctype html><textarea>a&lt;b</textarea>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "a<b"

#data
<!doctype html><iframe><!--<iframe></iframe>--></iframe>
#errors
(1,56): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <iframe>
|       "<!--<iframe>"
|     "-->"

#data
<!doctype html><iframe>...<!--X->...<!--/X->...</iframe>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <iframe>
|       "...<!--X->...<!--/X->..."

#data
<!doctype html><xmp><!--<xmp></xmp>--></xmp>
#errors
(1,44): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <xmp>
|       "<!--<xmp>"
|     "-->"

#data
<!doctype html><noembed><!--<noembed></noembed>--></noembed>
#errors
(1,60): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <noembed>
|       "<!--<noembed>"
|     "-->"

#data
<script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,8): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|   <body>

#data
<script>a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,9): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "a"
|   <body>

#data
<script><
#errors
(1,8): expected-doctype-but-got-start-tag
(1,9): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<"
|   <body>

#data
<script></
#errors
(1,8): expected-doctype-but-got-start-tag
(1,10): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</"
|   <body>

#data
<script></S
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</S"
|   <body>

#data
<script></SC
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SC"
|   <body>

#data
<script></SCR
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SCR"
|   <body>

#data
<script></SCRI
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SCRI"
|   <body>

#data
<script></SCRIP
#errors
(1,8): expected-doctype-but-got-start-tag
(1,15): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SCRIP"
|   <body>

#data
<script></SCRIPT
#errors
(1,8): expected-doctype-but-got-start-tag
(1,16): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</SCRIPT"
|   <body>

#data
<script></SCRIPT 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,17): expected-attribute-name-but-got-eof
(1,17): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|   <body>

#data
<script></s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</s"
|   <body>

#data
<script></sc
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</sc"
|   <body>

#data
<script></scr
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</scr"
|   <body>

#data
<script></scri
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</scri"
|   <body>

#data
<script></scrip
#errors
(1,8): expected-doctype-but-got-start-tag
(1,15): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</scrip"
|   <body>

#data
<script></script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,16): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "</script"
|   <body>

#data
<script></script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,17): expected-attribute-name-but-got-eof
(1,17): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|   <body>

#data
<script><!
#errors
(1,8): expected-doctype-but-got-start-tag
(1,10): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!"
|   <body>

#data
<script><!a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!a"
|   <body>

#data
<script><!-
#errors
(1,8): expected-doctype-but-got-start-tag
(1,11): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!-"
|   <body>

#data
<script><!-a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!-a"
|   <body>

#data
<script><!--
#errors
(1,8): expected-doctype-but-got-start-tag
(1,12): expected-named-closing-tag-but-got-eof
(1,12): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--"
|   <body>

#data
<script><!--a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
(1,13): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--a"
|   <body>

#data
<script><!--<
#errors
(1,8): expected-doctype-but-got-start-tag
(1,13): expected-named-closing-tag-but-got-eof
(1,13): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<"
|   <body>

#data
<script><!--<a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<a"
|   <body>

#data
<script><!--</
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--</"
|   <body>

#data
<script><!--</script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,20): expected-named-closing-tag-but-got-eof
(1,20): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--</script"
|   <body>

#data
<script><!--</script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): expected-attribute-name-but-got-eof
(1,21): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--"
|   <body>

#data
<script><!--<s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,14): expected-named-closing-tag-but-got-eof
(1,14): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<s"
|   <body>

#data
<script><!--<script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,19): expected-named-closing-tag-but-got-eof
(1,19): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script"
|   <body>

#data
<script><!--<script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,20): eof-in-script-in-script
(1,20): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script "
|   <body>

#data
<script><!--<script <
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): eof-in-script-in-script
(1,21): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script <"
|   <body>

#data
<script><!--<script <a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script <a"
|   <body>

#data
<script><!--<script </
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </"
|   <body>

#data
<script><!--<script </s
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): eof-in-script-in-script
(1,23): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </s"
|   <body>

#data
<script><!--<script </script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,28): eof-in-script-in-script
(1,28): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script"
|   <body>

#data
<script><!--<script </scripta
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): eof-in-script-in-script
(1,29): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </scripta"
|   <body>

#data
<script><!--<script </script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<script><!--<script </script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script>"
|   <body>

#data
<script><!--<script </script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,29): expected-named-closing-tag-but-got-eof
(1,29): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script/"
|   <body>

#data
<script><!--<script </script <
#errors
(1,8): expected-doctype-but-got-start-tag
(1,30): expected-named-closing-tag-but-got-eof
(1,30): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script <"
|   <body>

#data
<script><!--<script </script <a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
(1,31): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script <a"
|   <body>

#data
<script><!--<script </script </
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
(1,31): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script </"
|   <body>

#data
<script><!--<script </script </script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,37): expected-named-closing-tag-but-got-eof
(1,37): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script </script"
|   <body>

#data
<script><!--<script </script </script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,38): expected-attribute-name-but-got-eof
(1,38): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<script><!--<script </script </script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,38): unexpected-EOF-after-solidus-in-tag
(1,38): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<script><!--<script </script </script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script </script "
|   <body>

#data
<script><!--<script -
#errors
(1,8): expected-doctype-but-got-start-tag
(1,21): eof-in-script-in-script
(1,21): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -"
|   <body>

#data
<script><!--<script -a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -a"
|   <body>

#data
<script><!--<script --
#errors
(1,8): expected-doctype-but-got-start-tag
(1,22): eof-in-script-in-script
(1,22): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --"
|   <body>

#data
<script><!--<script --a
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): eof-in-script-in-script
(1,23): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --a"
|   <body>

#data
<script><!--<script -->
#errors
(1,8): expected-doctype-but-got-start-tag
(1,23): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<script><!--<script --><
#errors
(1,8): expected-doctype-but-got-start-tag
(1,24): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --><"
|   <body>

#data
<script><!--<script --></
#errors
(1,8): expected-doctype-but-got-start-tag
(1,25): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --></"
|   <body>

#data
<script><!--<script --></script
#errors
(1,8): expected-doctype-but-got-start-tag
(1,31): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script --></script"
|   <body>

#data
<script><!--<script --></script 
#errors
(1,8): expected-doctype-but-got-start-tag
(1,32): expected-attribute-name-but-got-eof
(1,32): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<script><!--<script --></script/
#errors
(1,8): expected-doctype-but-got-start-tag
(1,32): unexpected-EOF-after-solidus-in-tag
(1,32): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<script><!--<script --></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script -->"
|   <body>

#data
<script><!--<script><\/script>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script><\/script>-->"
|   <body>

#data
<script><!--<script></scr'+'ipt>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></scr'+'ipt>-->"
|   <body>

#data
<script><!--<script></script><script></script></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>"
|   <body>

#data
<script><!--<script></script><script></script>--><!--</script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>--><!--"
|   <body>

#data
<script><!--<script></script><script></script>-- ></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>-- >"
|   <body>

#data
<script><!--<script></script><script></script>- -></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>- ->"
|   <body>

#data
<script><!--<script></script><script></script>- - ></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>- - >"
|   <body>

#data
<script><!--<script></script><script></script>-></script>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></script><script></script>->"
|   <body>

#data
<script><!--<script>--!></script>X
#errors
(1,8): expected-doctype-but-got-start-tag
(1,34): expected-named-closing-tag-but-got-eof
(1,34): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script>--!></script>X"
|   <body>

#data
<script><!--<scr'+'ipt></script>--></script>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,44): unexpected-end-tag
#document
| <html>
|   <head>
|     <script>
|       "<!--<scr'+'ipt>"
|   <body>
|     "-->"

#data
<script><!--<script></scr'+'ipt></script>X
#errors
(1,8): expected-doctype-but-got-start-tag
(1,42): expected-named-closing-tag-but-got-eof
(1,42): unexpected-eof-in-text-mode
#document
| <html>
|   <head>
|     <script>
|       "<!--<script></scr'+'ipt></script>X"
|   <body>

#data
<style><!--<style></style>--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--<style>"
|   <body>
|     "-->"

#data
<style><!--</style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--"
|   <body>
|     "X"

#data
<style><!--...</style>...--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,36): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--..."
|   <body>
|     "...-->"

#data
<style><!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style></style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--<br><html xmlns:v="urn:schemas-microsoft-com:vml"><!--[if !mso]><style>"
|   <body>
|     "X"

#data
<style><!--...<style><!--...--!></style>--></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,51): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--...<style><!--...--!>"
|   <body>
|     "-->"

#data
<style><!--...</style><!-- --><style>@import ...</style>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--..."
|     <!--   -->
|     <style>
|       "@import ..."
|   <body>

#data
<style>...<style><!--...</style><!-- --></style>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,48): unexpected-end-tag
#document
| <html>
|   <head>
|     <style>
|       "...<style><!--..."
|     <!--   -->
|   <body>

#data
<style>...<!--[if IE]><style>...</style>X
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "...<!--[if IE]><style>..."
|   <body>
|     "X"

#data
<title><!--<title></title>--></title>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): unexpected-end-tag
#document
| <html>
|   <head>
|     <title>
|       "<!--<title>"
|   <body>
|     "-->"

#data
<title>&lt;/title></title>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <title>
|       "</title>"
|   <body>

#data
<title>foo/title><link></head><body>X
#errors
(1,7): expected-doctype-but-got-start-tag
(1,37): expected-named-closing-tag-but-got-eof
#document
| <html>
|   <head>
|     <title>
|       "foo/title><link></head><body>X"
|   <body>

#data
<noscript><!--<noscript></noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#script-on
#document
| <html>
|   <head>
|     <noscript>
|       "<!--<noscript>"
|   <body>
|     "-->"

#data
<noscript><!--<noscript></noscript>--></noscript>
#errors
 * (1,11) missing DOCTYPE
#script-off
#document
| <html>
|   <head>
|     <noscript>
|       <!-- <noscript></noscript> -->
|   <body>

#data
<noscript><!--</noscript>X<noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
#script-on
#document
| <html>
|   <head>
|     <noscript>
|       "<!--"
|   <body>
|     "X"
|     <noscript>
|       "-->"

#data
<noscript><!--</noscript>X<noscript>--></noscript>
#errors
(1,10): expected-doctype-but-got-start-tag
#script-off
#document
| <html>
|   <head>
|     <noscript>
|       <!-- </noscript>X<noscript> -->
|   <body>

#data
<noscript><iframe></noscript>X
#errors
(1,10): expected-doctype-but-got-start-tag
#script-on
#document
| <html>
|   <head>
|     <noscript>
|       "<iframe>"
|   <body>
|     "X"

#data
<noscript><iframe></noscript>X
#errors
 * (1,11) missing DOCTYPE
 * (1,19) unexpected token in head noscript
 * (1,31) unexpected EOF
#script-off
#document
| <html>
|   <head>
|     <noscript>
|   <body>
|     <iframe>
|       "</noscript>X"

#data
<noframes><!--<noframes></noframes>--></noframes>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#document
| <html>
|   <head>
|     <noframes>
|       "<!--<noframes>"
|   <body>
|     "-->"

#data
<noframes><body><script><!--...</script></body></noframes></html>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <noframes>
|       "<body><script><!--...</script></body>"
|   <body>

#data
<textarea><!--<textarea></textarea>--></textarea>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,49): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<!--<textarea>"
|     "-->"

#data
<textarea>&lt;/textarea></textarea>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "</textarea>"

#data
<iframe><!--<iframe></iframe>--></iframe>
#errors
(1,8): expected-doctype-but-got-start-tag
(1,41): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <iframe>
|       "<!--<iframe>"
|     "-->"

#data
<iframe>...<!--X->...<!--/X->...</iframe>
#errors
(1,8): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <iframe>
|       "...<!--X->...<!--/X->..."

#data
<xmp><!--<xmp></xmp>--></xmp>
#errors
(1,5): expected-doctype-but-got-start-tag
(1,29): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <xmp>
|       "<!--<xmp>"
|     "-->"

#data
<noembed><!--<noembed></noembed>--></noembed>
#errors
(1,9): expected-doctype-but-got-start-tag
(1,45): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     <noembed>
|       "<!--<noembed>"
|     "-->"

#data
<!doctype html><table>

#errors
(2,0): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "
"

#data
<!doctype html><table><td><span><font></span><span>
#errors
(1,26): unexpected-cell-in-table-body
(1,45): unexpected-end-tag
(1,51): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <span>
|               <font>
|             <font>
|               <span>

#data
<!doctype html><form><table></form><form></table></form>
#errors
(1,35): unexpected-end-tag-implies-table-voodoo
(1,35): unexpected-end-tag
(1,41): unexpected-form-in-table
(1,56): unexpected-end-tag
(1,56): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <form>
|       <table>
|         <form>

#data
<!doctype html><table><tbody><select><tr>
#errors
(1,37): unexpected-start-tag-implies-table-voodoo
(1,41): unexpected-table-element-start-tag-in-select-in-table
(1,41): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|       <tbody>
|         <tr>

#data
<!doctype html><table><tr><select><td>
#errors
(1,34): unexpected-start-tag-implies-table-voodoo
(1,38): unexpected-table-element-start-tag-in-select-in-table
(1,38): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!doctype html><table><tr><td><select><td>
#errors
(1,42): unexpected-table-element-start-tag-in-select-in-table
(1,42): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|           <td>

#data
<!doctype html><table><tr><th><select><td>
#errors
(1,42): unexpected-table-element-start-tag-in-select-in-table
(1,42): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <select>
|           <td>

#data
<!doctype html><table><caption><select><tr>
#errors
(1,43): unexpected-table-element-start-tag-in-select-in-table
(1,43): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <select>
|       <tbody>
|         <tr>

#data
<!doctype html><select><tr>
#errors
(1,27): unexpected-start-tag-in-select
(1,27): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><td>
#errors
(1,27): unexpected-start-tag-in-select
(1,27): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><th>
#errors
(1,27): unexpected-start-tag-in-select
(1,27): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><tbody>
#errors
(1,30): unexpected-start-tag-in-select
(1,30): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><thead>
#errors
(1,30): unexpected-start-tag-in-select
(1,30): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><tfoot>
#errors
(1,30): unexpected-start-tag-in-select
(1,30): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><caption>
#errors
(1,32): unexpected-start-tag-in-select
(1,32): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><table><tr></table>a
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|     "a"

#data
<!doctype html><plaintext></plaintext>
#errors
(1,38): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "</plaintext>"

#data
<!doctype html><table><plaintext></plaintext>
#errors
(1,33): foster-parenting-start-tag
(1,34): foster-parenting-character
(1,35): foster-parenting-character
(1,36): foster-parenting-character
(1,37): foster-parenting-character
(1,38): foster-parenting-character
(1,39): foster-parenting-character
(1,40): foster-parenting-character
(1,41): foster-parenting-character
(1,42): foster-parenting-character
(1,43): foster-parenting-character
(1,44): foster-parenting-character
(1,45): foster-parenting-character
(1,45): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "</plaintext>"
|     <table>

#data
<!doctype html><table><tbody><plaintext></plaintext>
#errors
(1,40): foster-parenting-start-tag
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,41): foster-parenting-character
(1,52): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "</plaintext>"
|     <table>
|       <tbody>

#data
<!doctype html><table><tbody><tr><plaintext></plaintext>
#errors
(1,44): foster-parenting-start-tag
(1,45): foster-parenting-character
(1,46): foster-parenting-character
(1,47): foster-parenting-character
(1,48): foster-parenting-character
(1,49): foster-parenting-character
(1,50): foster-parenting-character
(1,51): foster-parenting-character
(1,52): foster-parenting-character
(1,53): foster-parenting-character
(1,54): foster-parenting-character
(1,55): foster-parenting-character
(1,56): foster-parenting-character
(1,56): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <plaintext>
|       "</plaintext>"
|     <table>
|       <tbody>
|         <tr>

#data
<!doctype html><table><td><plaintext></plaintext>
#errors
(1,26): unexpected-cell-in-table-body
(1,49): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <plaintext>
|               "</plaintext>"

#data
<!doctype html><table><caption><plaintext></plaintext>
#errors
(1,54): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <plaintext>
|           "</plaintext>"

#data
<!doctype html><table><tr><style></script></style>abc
#errors
(1,51): foster-parenting-character
(1,52): foster-parenting-character
(1,53): foster-parenting-character
(1,53): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "abc"
|     <table>
|       <tbody>
|         <tr>
|           <style>
|             "</script>"

#data
<!doctype html><table><tr><script></style></script>abc
#errors
(1,52): foster-parenting-character
(1,53): foster-parenting-character
(1,54): foster-parenting-character
(1,54): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "abc"
|     <table>
|       <tbody>
|         <tr>
|           <script>
|             "</style>"

#data
<!doctype html><table><caption><style></script></style>abc
#errors
(1,58): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <style>
|           "</script>"
|         "abc"

#data
<!doctype html><table><td><style></script></style>abc
#errors
(1,26): unexpected-cell-in-table-body
(1,53): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <style>
|               "</script>"
|             "abc"

#data
<!doctype html><select><script></style></script>abc
#errors
(1,51): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <script>
|         "</style>"
|       "abc"

#data
<!doctype html><table><select><script></style></script>abc
#errors
(1,30): unexpected-start-tag-implies-table-voodoo
(1,58): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <script>
|         "</style>"
|       "abc"
|     <table>

#data
<!doctype html><table><tr><select><script></style></script>abc
#errors
(1,34): unexpected-start-tag-implies-table-voodoo
(1,62): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <script>
|         "</style>"
|       "abc"
|     <table>
|       <tbody>
|         <tr>

#data
<!doctype html><frameset></frameset><noframes>abc
#errors
(1,49): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abc"

#data
<!doctype html><frameset></frameset><noframes>abc</noframes><!--abc-->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abc"
|   <!-- abc -->

#data
<!doctype html><frameset></frameset></html><noframes>abc
#errors
(1,56): expected-named-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abc"

#data
<!doctype html><frameset></frameset></html><noframes>abc</noframes><!--abc-->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abc"
| <!-- abc -->

#data
<!doctype html><table><tr></tbody><tfoot>
#errors
(1,41): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|       <tfoot>

#data
<!doctype html><table><td><svg></svg>abc<td>
#errors
(1,26): unexpected-cell-in-table-body
(1,44): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|             "abc"
|           <td>

#data
<!doctype html><math><mn DefinitionUrl="foo">
#errors
(1,45): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mn>
|         definitionURL="foo"

#data
<!doctype html><html></p><!--foo-->
#errors
(1,25): end-tag-after-implied-root
#document
| <!DOCTYPE html>
| <html>
|   <!-- foo -->
|   <head>
|   <body>

#data
<!doctype html><head></head></p><!--foo-->
#errors
(1,32): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <!-- foo -->
|   <body>

#data
<!doctype html><body><p><pre>
#errors
(1,29): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <pre>

#data
<!doctype html><body><p><listing>
#errors
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <listing>

#data
<!doctype html><p><plaintext>
#errors
(1,29): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <plaintext>

#data
<!doctype html><p><h1>
#errors
(1,22): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <h1>

#data
<!doctype html><isindex type="hidden">
#errors
(1,38): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <isindex>
|       type="hidden"

#data
<!doctype html><ruby><div><span><rp>
#errors
(1,36): XXX-undefined-error
(1,36): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <div>
|         <span>
|           <rp>

#data
<!doctype html><ruby><div><p><rp>
#errors
(1,33): XXX-undefined-error
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <div>
|         <p>
|         <rp>

#data
<!doctype html><ruby><p><rt>
#errors
(1,28): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <p>
|       <rt>

#data
<!doctype html><ruby><div><span><rt>
#errors
(1,36): XXX-undefined-error
(1,36): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <div>
|         <span>
|           <rt>

#data
<!doctype html><ruby><div><p><rt>
#errors
(1,33): XXX-undefined-error
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <div>
|         <p>
|         <rt>

#data
<html><ruby>a<rb>b<rt></ruby></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rb>
|         "b"
|       <rt>

#data
<html><ruby>a<rp>b<rt></ruby></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rp>
|         "b"
|       <rt>

#data
<html><ruby>a<rt>b<rt></ruby></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rt>
|         "b"
|       <rt>

#data
<html><ruby>a<rtc>b<rt>c<rb>d</ruby></html>
#errors
(1,6): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rtc>
|         "b"
|         <rt>
|           "c"
|       <rb>
|         "d"

#data
<!doctype html><math/><foo>
#errors
(1,27): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|     <foo>

#data
<!doctype html><svg/><foo>
#errors
(1,26): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <foo>

#data
<!doctype html><div></body><!--foo-->
#errors
(1,27): expected-one-end-tag-but-got-another
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>
|   <!-- foo -->

#data
<!doctype html><h1><div><h3><span></h1>foo
#errors
(1,39): end-tag-too-early
(1,42): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <h1>
|       <div>
|         <h3>
|           <span>
|         "foo"

#data
<!doctype html><p></h3>foo
#errors
(1,23): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "foo"

#data
<!doctype html><h3><li>abc</h2>foo
#errors
(1,31): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <h3>
|       <li>
|         "abc"
|     "foo"

#data
<!doctype html><table>abc<!--foo-->
#errors
(1,23): foster-parenting-character
(1,24): foster-parenting-character
(1,25): foster-parenting-character
(1,35): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "abc"
|     <table>
|       <!-- foo -->

#data
<!doctype html><table>  <!--foo-->
#errors
(1,34): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "  "
|       <!-- foo -->

#data
<!doctype html><table> b <!--foo-->
#errors
(1,23): foster-parenting-character
(1,24): foster-parenting-character
(1,25): foster-parenting-character
(1,35): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " b "
|     <table>
|       <!-- foo -->

#data
<!doctype html><select><option><option>
#errors
(1,39): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <option>

#data
<!doctype html><select><option></optgroup>
#errors
(1,42): unexpected-end-tag-in-select
(1,42): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<!doctype html><select><option></optgroup>
#errors
(1,42): unexpected-end-tag-in-select
(1,42): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<!doctype html><dd><optgroup><dd>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <dd>
|       <optgroup>
|     <dd>

#data
<!doctype html><p><math><mi><p><h1>
#errors
(1,35): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math mi>
|           <p>
|           <h1>

#data
<!doctype html><p><math><mo><p><h1>
#errors
(1,35): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math mo>
|           <p>
|           <h1>

#data
<!doctype html><p><math><mn><p><h1>
#errors
(1,35): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math mn>
|           <p>
|           <h1>

#data
<!doctype html><p><math><ms><p><h1>
#errors
(1,35): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math ms>
|           <p>
|           <h1>

#data
<!doctype html><p><math><mtext><p><h1>
#errors
(1,38): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math mtext>
|           <p>
|           <h1>

#data
<!doctype html><frameset></noframes>
#errors
(1,36): unexpected-end-tag-in-frameset
(1,36): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!doctype html><html c=d><body></html><html a=b>
#errors
(1,48): non-html-root
#document
| <!DOCTYPE html>
| <html>
|   a="b"
|   c="d"
|   <head>
|   <body>

#data
<!doctype html><html c=d><frameset></frameset></html><html a=b>
#errors
(1,63): non-html-root
#document
| <!DOCTYPE html>
| <html>
|   a="b"
|   c="d"
|   <head>
|   <frameset>

#data
<!doctype html><html><frameset></frameset></html><!--foo-->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
| <!-- foo -->

#data
<!doctype html><html><frameset></frameset></html>  
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   "  "

#data
<!doctype html><html><frameset></frameset></html>abc
#errors
(1,50): expected-eof-but-got-char
(1,51): expected-eof-but-got-char
(1,52): expected-eof-but-got-char
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!doctype html><html><frameset></frameset></html><p>
#errors
(1,52): expected-eof-but-got-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!doctype html><html><frameset></frameset></html></p>
#errors
(1,53): expected-eof-but-got-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<html><frameset></frameset></html><!doctype html>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,49): unexpected-doctype
#document
| <html>
|   <head>
|   <frameset>

#data
<!doctype html><body><frameset>
#errors
(1,31): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!doctype html><p><frameset><frame>
#errors
(1,28): unexpected-start-tag
(1,35): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|     <frame>

#data
<!doctype html><p>a<frameset>
#errors
(1,29): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "a"

#data
<!doctype html><p> <frameset><frame>
#errors
(1,29): unexpected-start-tag
(1,36): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|     <frame>

#data
<!doctype html><pre><frameset>
#errors
(1,30): unexpected-start-tag
(1,30): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <pre>

#data
<!doctype html><listing><frameset>
#errors
(1,34): unexpected-start-tag
(1,34): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <listing>

#data
<!doctype html><li><frameset>
#errors
(1,29): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <li>

#data
<!doctype html><dd><frameset>
#errors
(1,29): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <dd>

#data
<!doctype html><dt><frameset>
#errors
(1,29): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <dt>

#data
<!doctype html><button><frameset>
#errors
(1,33): unexpected-start-tag
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <button>

#data
<!doctype html><applet><frameset>
#errors
(1,33): unexpected-start-tag
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <applet>

#data
<!doctype html><marquee><frameset>
#errors
(1,34): unexpected-start-tag
(1,34): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <marquee>

#data
<!doctype html><object><frameset>
#errors
(1,33): unexpected-start-tag
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <object>

#data
<!doctype html><table><frameset>
#errors
(1,32): unexpected-start-tag-implies-table-voodoo
(1,32): unexpected-start-tag
(1,32): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>

#data
<!doctype html><area><frameset>
#errors
(1,31): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <area>

#data
<!doctype html><basefont><frameset>
#errors
(1,35): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <basefont>
|   <frameset>

#data
<!doctype html><bgsound><frameset>
#errors
(1,34): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <bgsound>
|   <frameset>

#data
<!doctype html><br><frameset>
#errors
(1,29): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <br>

#data
<!doctype html><embed><frameset>
#errors
(1,32): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <embed>

#data
<!doctype html><img><frameset>
#errors
(1,30): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <img>

#data
<!doctype html><input><frameset>
#errors
(1,32): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <input>

#data
<!doctype html><keygen><frameset>
#errors
(1,33): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <keygen>

#data
<!doctype html><wbr><frameset>
#errors
(1,30): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <wbr>

#data
<!doctype html><hr><frameset>
#errors
(1,29): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <hr>

#data
<!doctype html><textarea></textarea><frameset>
#errors
(1,46): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>

#data
<!doctype html><xmp></xmp><frameset>
#errors
(1,36): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!doctype html><iframe></iframe><frameset>
#errors
(1,42): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!doctype html><select></select><frameset>
#errors
(1,42): unexpected-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><svg></svg><frameset><frame>
#errors
(1,36): unexpected-start-tag
(1,43): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|     <frame>

#data
<!doctype html><math></math><frameset><frame>
#errors
(1,38): unexpected-start-tag
(1,45): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|     <frame>

#data
<!doctype html><svg><foreignObject><div> <frameset><frame>
#errors
(1,51): unexpected-start-tag
(1,58): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|     <frame>

#data
<!doctype html><svg>a</svg><frameset><frame>
#errors
(1,37): unexpected-start-tag
(1,44): unexpected-start-tag-ignored
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       "a"

#data
<!doctype html><svg> </svg><frameset><frame>
#errors
(1,37): unexpected-start-tag
(1,44): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|     <frame>

#data
<html>aaa<frameset></frameset>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,19): unexpected-start-tag
(1,30): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     "aaa"

#data
<html> a <frameset></frameset>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,19): unexpected-start-tag
(1,30): unexpected-end-tag
#document
| <html>
|   <head>
|   <body>
|     "a "

#data
<!doctype html><div><frameset>
#errors
(1,30): unexpected-start-tag
(1,30): eof-in-frameset
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!doctype html><div><body><frameset>
#errors
(1,26): unexpected-start-tag
(1,36): unexpected-start-tag
(1,36): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>

#data
<!doctype html><p><math></p>a
#errors
(1,28): unexpected-end-tag
(1,28): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|     "a"

#data
<!doctype html><p><math><mn><span></p>a
#errors
(1,38): unexpected-end-tag
(1,39): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <math math>
|         <math mn>
|           <span>
|             <p>
|             "a"

#data
<!doctype html><math></html>
#errors
(1,28): unexpected-end-tag
(1,28): expected-one-end-tag-but-got-another
(1,28): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>

#data
<!doctype html><meta charset="ascii">
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <meta>
|       charset="ascii"
|   <body>

#data
<!doctype html><meta http-equiv="content-type" content="text/html;charset=ascii">
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <meta>
|       content="text/html;charset=ascii"
|       http-equiv="content-type"
|   <body>

#data
<!doctype html><head><!--aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa--><meta charset="utf8">
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <!-- aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa -->
|     <meta>
|       charset="utf8"
|   <body>

#data
<!doctype html><html a=b><head></head><html c=d>
#errors
(1,48): non-html-root
#document
| <!DOCTYPE html>
| <html>
|   a="b"
|   c="d"
|   <head>
|   <body>

#data
<!doctype html><image/>
#errors
(1,23): image-start-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <img>

#data
<!doctype html>a<i>b<table>c<b>d</i>e</b>f
#errors
(1,28): foster-parenting-character
(1,31): foster-parenting-start-tag
(1,32): foster-parenting-character
(1,36): foster-parenting-end-tag
(1,36): adoption-agency-1.3
(1,37): foster-parenting-character
(1,41): foster-parenting-end-tag
(1,42): foster-parenting-character
(1,42): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a"
|     <i>
|       "bc"
|       <b>
|         "de"
|       "f"
|       <table>

#data
<!doctype html><table><i>a<b>b<div>c<a>d</i>e</b>f
#errors
(1,25): foster-parenting-start-tag
(1,26): foster-parenting-character
(1,29): foster-parenting-start-tag
(1,30): foster-parenting-character
(1,35): foster-parenting-start-tag
(1,36): foster-parenting-character
(1,39): foster-parenting-start-tag
(1,40): foster-parenting-character
(1,44): foster-parenting-end-tag
(1,44): adoption-agency-1.3
(1,44): adoption-agency-1.3
(1,45): foster-parenting-character
(1,49): foster-parenting-end-tag
(1,49): adoption-agency-1.3
(1,49): adoption-agency-1.3
(1,50): foster-parenting-character
(1,50): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <b>
|         "b"
|     <b>
|     <div>
|       <b>
|         <i>
|           "c"
|           <a>
|             "d"
|         <a>
|           "e"
|       <a>
|         "f"
|     <table>

#data
<!doctype html><i>a<b>b<div>c<a>d</i>e</b>f
#errors
(1,37): adoption-agency-1.3
(1,37): adoption-agency-1.3
(1,42): adoption-agency-1.3
(1,42): adoption-agency-1.3
(1,43): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <b>
|         "b"
|     <b>
|     <div>
|       <b>
|         <i>
|           "c"
|           <a>
|             "d"
|         <a>
|           "e"
|       <a>
|         "f"

#data
<!doctype html><table><i>a<b>b<div>c</i>
#errors
(1,25): foster-parenting-start-tag
(1,26): foster-parenting-character
(1,29): foster-parenting-start-tag
(1,30): foster-parenting-character
(1,35): foster-parenting-start-tag
(1,36): foster-parenting-character
(1,40): foster-parenting-end-tag
(1,40): adoption-agency-1.3
(1,40): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <b>
|         "b"
|     <b>
|       <div>
|         <i>
|           "c"
|     <table>

#data
<!doctype html><table><i>a<b>b<div>c<a>d</i>e</b>f
#errors
(1,25): foster-parenting-start-tag
(1,26): foster-parenting-character
(1,29): foster-parenting-start-tag
(1,30): foster-parenting-character
(1,35): foster-parenting-start-tag
(1,36): foster-parenting-character
(1,39): foster-parenting-start-tag
(1,40): foster-parenting-character
(1,44): foster-parenting-end-tag
(1,44): adoption-agency-1.3
(1,44): adoption-agency-1.3
(1,45): foster-parenting-character
(1,49): foster-parenting-end-tag
(1,44): adoption-agency-1.3
(1,44): adoption-agency-1.3
(1,50): foster-parenting-character
(1,50): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <b>
|         "b"
|     <b>
|     <div>
|       <b>
|         <i>
|           "c"
|           <a>
|             "d"
|         <a>
|           "e"
|       <a>
|         "f"
|     <table>

#data
<!doctype html><table><i>a<div>b<tr>c<b>d</i>e
#errors
(1,25): foster-parenting-start-tag
(1,26): foster-parenting-character
(1,31): foster-parenting-start-tag
(1,32): foster-parenting-character
(1,37): foster-parenting-character
(1,40): foster-parenting-start-tag
(1,41): foster-parenting-character
(1,45): foster-parenting-end-tag
(1,45): adoption-agency-1.3
(1,46): foster-parenting-character
(1,46): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <i>
|       "a"
|       <div>
|         "b"
|     <i>
|       "c"
|       <b>
|         "d"
|     <b>
|       "e"
|     <table>
|       <tbody>
|         <tr>

#data
<!doctype html><table><td><table><i>a<div>b<b>c</i>d
#errors
(1,26): unexpected-cell-in-table-body
(1,36): foster-parenting-start-tag
(1,37): foster-parenting-character
(1,42): foster-parenting-start-tag
(1,43): foster-parenting-character
(1,46): foster-parenting-start-tag
(1,47): foster-parenting-character
(1,51): foster-parenting-end-tag
(1,51): adoption-agency-1.3
(1,51): adoption-agency-1.3
(1,52): foster-parenting-character
(1,52): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <i>
|               "a"
|             <div>
|               <i>
|                 "b"
|                 <b>
|                   "c"
|               <b>
|                 "d"
|             <table>

#data
<!doctype html><body><bgsound>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <bgsound>

#data
<!doctype html><body><basefont>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <basefont>

#data
<!doctype html><a><b></a><basefont>
#errors
(1,25): adoption-agency-1.3
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|     <basefont>

#data
<!doctype html><a><b></a><bgsound>
#errors
(1,25): adoption-agency-1.3
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|     <bgsound>

#data
<!doctype html><figcaption><article></figcaption>a
#errors
(1,49): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <figcaption>
|       <article>
|     "a"

#data
<!doctype html><summary><article></summary>a
#errors
(1,43): end-tag-too-early
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <summary>
|       <article>
|     "a"

#data
<!doctype html><p><a><plaintext>b
#errors
(1,32): unexpected-end-tag
(1,33): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <a>
|     <plaintext>
|       <a>
|         "b"

#data
<!DOCTYPE html><div>a<a></div>b<p>c</p>d
#errors
(1,30): end-tag-too-early
(1,40): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <div>
|       "a"
|       <a>
|     <a>
|       "b"
|       <p>
|         "c"
|       "d"
